Calculate age/sex/area stratified survey estimates for biomarker outcomes

Calculate age/sex/area stratified survey estimates for biomarker outcomes

calc_survey_hiv_indicators(survey_meta, survey_regions, survey_clusters,
  survey_individuals, survey_biomarker, areas, sex = c("male", "female",
  "both"), age_group_id = NULL, area_top_level = min(areas$area_level),
  area_bottom_level = max(areas$area_level))

Arguments

survey_meta
Survey metadata.

survey_regions
Survey regions.

survey_clusters
Survey clusters.

survey_individuals
Survey individuals.

survey_biomarker
Survey biomarkers.

areas
Areas.

sex
Sex.

age_group_id
Age group id.

area_top_level
Area top level.

area_bottom_level
Area bottom level.

Details

All other data will be subsetted based on the survey_id values appearing in survey_meta, so if only want to calculate for a subset of surveys it is sufficient to pass subset for survey_meta and full data frames for the others.

Much of this function needs to be parsed out into more generic functions and rewritten to be more efficient.

Age group would be more efficient if traversing a tree structure.

Need generic function to calculate

Flexibility about age/sex stratifications to calculate.