Programme data: Malawi routine HIV service delivery data

Datasets containing Malawi National HIV programme quarterly reported data about:

Reported number currently on ART.

Number of women receiving antenatal care and final antenatal HIV testing.

demo_anc_testing

demo_art_number

Format

TODO: Link to demo_areas dataset demo_art_number is a dataset with 2042 observations on the following 5 variables:

area_id

District area id – see demo_areas dataset.

sex

Sex as a character vector – all values are "both".

age_group

Value "Y015_999" is age 15+, "Y000_014" is age 0-14.

calendar_quarter

Calendar quarter.

art_current

Number currently on ART at the end of the quarter.

art_new

Number newly initiated ART during the quarter.

TODO: Link to demo_areas dataset demo_anc_testing is a dataset with 923 observations on the following 8 variables:

area_id

District area id – see demo_areas dataset.

quarter_id

Quarter ID (integer).

anc_clients

Number of unique ANC clients during quarter.

anc_known_pos

Number of ANC clients already known to be HIV positive prior to first ANC visit and not tested for HIV.

anc_already_art

Number of ANC clients already on ART prior to first ANC visit.

anc_tested

Number of ANC clients tested for HIV and received result.

anc_tested_pos

Number of ANC clients testing HIV postive.

anc_known_pos

Number of ANC clients not tested for HIV due to documented recent HIV negative status (optional).

births_facility

Number of life births recorded at health facilities..

An object of class tbl_df (inherits from tbl, data.frame) with 830 rows and 10 columns.

Source

Department of HIV & AIDS, Minitry of Health, Malawi. http://hiv.health.gov.mw/index.php/our-documents

Details

Data are sourced from quarterly reporting spreadsheets produced via the Quarterly Integrated HIV Program Reports and publicly available from Malawi Ministry of Health: Department of HIV & AIDS website.

Data are provided by the Ministry of Health, Malawi with the following disclaimer:

The data in this file are owned by the Ministry of Health (MOH), Malawi.

These data are shared in order to support the National HIV Program.

Data in this file may not be used by Third Parties for further analysis and dissemination without prior written approval of the Director of the Department for HIV and AIDS, MOH.

Patient-level data are routinely collected by health facility staff using standard monitoring tools. This file contains facility-level aggregates that have been generated by facility staff and that were verified from primary records during quarterly National HIV Program Supervision (coordinated of the Department of HIV and AIDS). While every effort is made to ensure high data quality, individual records may not be complete and accurate.

The official interpretation of these data is presented in Quarterly Integrated HIV Program Reports and any divergent interpretation may be misleading and is not supported by the MOH. Analysis and interpretation of the data requires detailed understsanding of the methods and constraints of the Malawi's National M&E System.

The number of ART by age 0-14 and age 15+ is not reported in aggregate quarterly reporting data. For model illustration purposes, this is approximated as 94% of all ART clients are age 15+ based on Spectrum model outputs which were validated against age distributions from electronic medical records by the Malawi HIV estimates team.

Dataset could contain one observation per district per quarter for all 32 districts for all quarters between Q3 2011 and Q3 2018, but the number of observations is fewer than 32 * 29 = 928 due to a few missing observations.

See also