Create district bar plot

Create district bar plot

district_barplot(
  naomi_output,
  indicator,
  age,
  sex_disag,
  level = NULL,
  calendar_quarter = 2,
  label_format,
  x_title = NULL,
  fig_title = NULL,
  legend_title = NULL,
  colour_palette = "blue",
  above_label = t_("ABOVE"),
  below_label = t_("BELOW")
)

Arguments

naomi_output

Naomi output object or path to naomi output zip folder

indicator

Indicator to plot

age

Age group to plot

sex_disag

Sexes to plot

level

Area level to plot, default all

calendar_quarter

Naomi projection quarter to filter data to, default is calendar_quarter_t2

label_format

Label format for y labels

x_title

Title for the figure x axis, blank by default

fig_title

Title for the figure, blank by default,

legend_title

Title for the legend

colour_palette

Names or hexcode value for colours to differentiate districts with values above and below the national average. Can be specified as a preset colour palette using "red", "green" and "blue" or manually as a named list: cols <- c(above_colour = "red", below_colour = "blue").

above_label

Label for districts with values higher than national average, default is "Above"

below_label

Label for districts with values higher than national average, default is "Below"

Value

District barplot