Export proportion aware by five year age group from Shiny90

Export estimates for proportion aware of status from a Shiny90 output file to five-year age groups 15-19 to 50+.

extract_shiny90_age_sex(shiny90_path, years = NULL)

Arguments

shiny90_path
file path to .shiny90 digest file.

years
year(s) to generate estimates; an integer or a vector of integers. If NULL, all years available in estimates are returned (default).

Value

A data frame consisting of the number of PLHIV, aware of status and on ART by sex and five year age group 15-49 and age 50+ for specified years.

Details

The 'artnum' divided by 'plhiv' columns in the output will give a different ART coverage than Spectrum output for the same year because these outputs are based on the internal mid-year ART coverage in Spectrum, while Spectrum reports an end of year (Dec 31) ART coverage.

Examples

pjnz <- system.file("extdata/demo_mwi2019.PJNZ", package = "naomi")
shiny90dir <- tempfile()
utils::unzip(pjnz, "malawi.zip.shiny90", exdir = shiny90dir)
shiny90_path <- file.path(shiny90dir, "malawi.zip.shiny90")

extract_shiny90_age_sex(shiny90_path, year = 2010:2019)

#>       area year    sex agegr hivstatus     plhiv      aware     artnum
#> 1   Malawi 2010 female 15-19  positive  27661.84  13519.155   3875.436
#> 2   Malawi 2011 female 15-19  positive  28170.73  14522.437   5450.108
#> 3   Malawi 2012 female 15-19  positive  28605.54  15531.726   7570.642
#> 4   Malawi 2013 female 15-19  positive  28900.06  15865.382   9587.347
#> 5   Malawi 2014 female 15-19  positive  29094.08  16507.558  11245.004
#> 6   Malawi 2015 female 15-19  positive  29224.79  17112.835  12527.553
#> 7   Malawi 2016 female 15-19  positive  29287.59  17908.189  14173.245
#> 8   Malawi 2017 female 15-19  positive  29099.98  19641.169  15963.250
#> 9   Malawi 2018 female 15-19  positive  28704.54  20664.685  17333.081
#> 10  Malawi 2019 female 15-19  positive  28120.45  20953.640  17964.586
#> 11  Malawi 2010   male 15-19  positive  11628.93   5971.575   2950.240
#> 12  Malawi 2011   male 15-19  positive  12655.30   6853.500   3866.402
#> 13  Malawi 2012   male 15-19  positive  13556.78   7638.029   4892.732
#> 14  Malawi 2013   male 15-19  positive  14326.91   8336.562   5913.852
#> 15  Malawi 2014   male 15-19  positive  14977.45   9142.929   6886.548
#> 16  Malawi 2015   male 15-19  positive  15545.93   9935.907   7819.709
#> 17  Malawi 2016   male 15-19  positive  16016.81  10710.855   8711.919
#> 18  Malawi 2017   male 15-19  positive  16333.26  11745.458   9749.499
#> 19  Malawi 2018   male 15-19  positive  16523.14  12667.399  10795.934
#> 20  Malawi 2019   male 15-19  positive  16611.00  13304.368  11230.149
#> 21  Malawi 2010 female 20-24  positive  55860.31  35145.428   6416.244
#> 22  Malawi 2011 female 20-24  positive  56349.19  37370.670   9636.857
#> 23  Malawi 2012 female 20-24  positive  56510.19  38981.897  14051.622
#> 24  Malawi 2013 female 20-24  positive  56240.97  39182.456  18380.011
#> 25  Malawi 2014 female 20-24  positive  55677.77  39536.217  21780.435
#> 26  Malawi 2015 female 20-24  positive  55020.53  39776.054  24352.369
#> 27  Malawi 2016 female 20-24  positive  54412.27  40397.180  27735.551
#> 28  Malawi 2017 female 20-24  positive  53736.27  42547.117  31518.472
#> 29  Malawi 2018 female 20-24  positive  52926.04  43690.424  34432.341
#> 30  Malawi 2019 female 20-24  positive  51909.37  43801.394  35888.204
#> 31  Malawi 2010   male 20-24  positive  19474.99   7862.534   1480.791
#> 32  Malawi 2011   male 20-24  positive  20488.62   8642.073   2129.633
#> 33  Malawi 2012   male 20-24  positive  21418.28   9346.702   2901.717
#> 34  Malawi 2013   male 20-24  positive  22143.16   9887.536   3707.623
#> 35  Malawi 2014   male 20-24  positive  22698.00  10592.273   4771.351
#> 36  Malawi 2015   male 20-24  positive  23148.76  11317.334   5747.568
#> 37  Malawi 2016   male 20-24  positive  23552.99  12201.725   7073.690
#> 38  Malawi 2017   male 20-24  positive  23841.47  13754.657   8693.387
#> 39  Malawi 2018   male 20-24  positive  23987.74  15064.006  10254.990
#> 40  Malawi 2019   male 20-24  positive  23980.94  15937.333  10840.859
#> 41  Malawi 2010 female 25-29  positive  73374.83  54350.522  14129.024
#> 42  Malawi 2011 female 25-29  positive  72708.45  56543.377  18311.125
#> 43  Malawi 2012 female 25-29  positive  72907.34  58497.997  24379.249
#> 44  Malawi 2013 female 25-29  positive  73799.11  59741.664  30506.387
#> 45  Malawi 2014 female 25-29  positive  75005.26  61467.069  35943.826
#> 46  Malawi 2015 female 25-29  positive  76056.00  63042.898  40508.478
#> 47  Malawi 2016 female 25-29  positive  76551.54  64544.780  45647.565
#> 48  Malawi 2017 female 25-29  positive  76192.44  66873.752  50887.326
#> 49  Malawi 2018 female 25-29  positive  75029.45  67591.496  54615.850
#> 50  Malawi 2019 female 25-29  positive  73313.28  66955.491  56188.160
#> 51  Malawi 2010   male 25-29  positive  37824.77  22230.649   3877.811
#> 52  Malawi 2011   male 25-29  positive  37426.88  22259.260   4784.630
#> 53  Malawi 2012   male 25-29  positive  37564.77  22259.695   5820.615
#> 54  Malawi 2013   male 25-29  positive  38112.84  22252.003   6825.445
#> 55  Malawi 2014   male 25-29  positive  38900.93  22868.188   8386.136
#> 56  Malawi 2015   male 25-29  positive  39697.06  23705.184   9843.344
#> 57  Malawi 2016   male 25-29  positive  40285.63  24871.801  11911.549
#> 58  Malawi 2017   male 25-29  positive  40432.99  27216.228  14548.514
#> 59  Malawi 2018   male 25-29  positive  40124.30  28880.542  17031.039
#> 60  Malawi 2019   male 25-29  positive  39473.38  29624.525  17668.173
#> 61  Malawi 2010 female 30-34  positive  87174.83  69531.953  22469.918
#> 62  Malawi 2011 female 30-34  positive  86025.51  72260.312  27843.411
#> 63  Malawi 2012 female 30-34  positive  84646.10  73438.968  34819.282
#> 64  Malawi 2013 female 30-34  positive  83274.51  73216.301  41017.257
#> 65  Malawi 2014 female 30-34  positive  82135.71  73036.061  46063.513
#> 66  Malawi 2015 female 30-34  positive  81466.92  73040.521  50135.703
#> 67  Malawi 2016 female 30-34  positive  81492.13  73806.710  55050.959
#> 68  Malawi 2017 female 30-34  positive  82174.45  76182.259  60770.683
#> 69  Malawi 2018 female 30-34  positive  83343.20  78465.744  65986.620
#> 70  Malawi 2019 female 30-34  positive  84576.71  80275.624  69731.825
#> 71  Malawi 2010   male 30-34  positive  57120.76  39923.596   8845.565
#> 72  Malawi 2011   male 30-34  positive  55587.50  39971.019  10554.645
#> 73  Malawi 2012   male 30-34  positive  54017.62  39185.502  12202.337
#> 74  Malawi 2013   male 30-34  positive  52531.33  37906.924  13462.247
#> 75  Malawi 2014   male 30-34  positive  51273.88  37059.075  15003.887
#> 76  Malawi 2015   male 30-34  positive  50410.68  36585.750  16361.271
#> 77  Malawi 2016   male 30-34  positive  50080.79  36866.429  18422.836
#> 78  Malawi 2017   male 30-34  positive  50236.51  38880.158  21521.342
#> 79  Malawi 2018   male 30-34  positive  50747.01  40986.065  24872.495
#> 80  Malawi 2019   male 30-34  positive  51354.07  42640.519  26257.567
#> 81  Malawi 2010 female 35-39  positive  80749.88  58867.751  27515.654
#> 82  Malawi 2011 female 35-39  positive  83229.06  66065.546  34747.218
#> 83  Malawi 2012 female 35-39  positive  85424.20  71768.257  43857.259
#> 84  Malawi 2013 female 35-39  positive  87201.02  75652.092  51976.999
#> 85  Malawi 2014 female 35-39  positive  88362.07  78435.552  58348.885
#> 86  Malawi 2015 female 35-39  positive  88805.76  80074.292  63229.217
#> 87  Malawi 2016 female 35-39  positive  88530.30  80901.472  67864.922
#> 88  Malawi 2017 female 35-39  positive  87662.10  81672.352  71972.434
#> 89  Malawi 2018 female 35-39  positive  86514.02  81816.218  74631.929
#> 90  Malawi 2019 female 35-39  positive  85415.51  81528.792  75703.008
#> 91  Malawi 2010   male 35-39  positive  62690.76  39932.267  14848.291
#> 92  Malawi 2011   male 35-39  positive  62779.79  42911.368  18011.952
#> 93  Malawi 2012   male 35-39  positive  62775.06  44776.082  21148.283
#> 94  Malawi 2013   male 35-39  positive  62620.65  45740.547  23702.641
#> 95  Malawi 2014   male 35-39  positive  62189.25  46261.108  25936.501
#> 96  Malawi 2015   male 35-39  positive  61420.95  46323.063  27756.438
#> 97  Malawi 2016   male 35-39  positive  60303.38  46177.651  29827.403
#> 98  Malawi 2017   male 35-39  positive  58916.54  46582.388  32611.991
#> 99  Malawi 2018   male 35-39  positive  57464.71  46942.898  35131.218
#> 100 Malawi 2019   male 35-39  positive  56163.52  47052.102  35270.767
#> 101 Malawi 2010 female 40-44  positive  57589.51  40645.608  22276.481
#> 102 Malawi 2011 female 40-44  positive  61689.76  47826.031  28784.209
#> 103 Malawi 2012 female 40-44  positive  65884.90  54567.722  37134.102
#> 104 Malawi 2013 female 40-44  positive  70107.88  60446.231  45282.302
#> 105 Malawi 2014 female 40-44  positive  74340.49  66049.984  52721.376
#> 106 Malawi 2015 female 40-44  positive  78875.55  71590.509  59948.856
#> 107 Malawi 2016 female 40-44  positive  82096.84  75826.316  66647.531
#> 108 Malawi 2017 female 40-44  positive  84755.49  79892.862  72956.880
#> 109 Malawi 2018 female 40-44  positive  86752.02  82990.509  77834.072
#> 110 Malawi 2019 female 40-44  positive  87983.73  84938.826  80703.447
#> 111 Malawi 2010   male 40-44  positive  51779.98  32404.574  15118.233
#> 112 Malawi 2011   male 40-44  positive  53818.81  36496.959  18946.960
#> 113 Malawi 2012   male 40-44  positive  55703.95  39914.333  22943.510
#> 114 Malawi 2013   male 40-44  positive  57411.38  42689.231  26577.940
#> 115 Malawi 2014   male 40-44  positive  58986.38  45228.474  29810.677
#> 116 Malawi 2015   male 40-44  positive  60848.50  47780.534  33107.447
#> 117 Malawi 2016   male 40-44  positive  61587.22  49436.632  36156.017
#> 118 Malawi 2017   male 40-44  positive  62009.97  51410.077  39887.498
#> 119 Malawi 2018   male 40-44  positive  62104.83  53015.763  43278.364
#> 120 Malawi 2019   male 40-44  positive  61838.02  53938.414  44111.280
#> 121 Malawi 2010 female 45-49  positive  35854.62  25098.386  15425.926
#> 122 Malawi 2011 female 45-49  positive  39457.67  30398.134  20091.560
#> 123 Malawi 2012 female 45-49  positive  43408.61  35820.025  26309.388
#> 124 Malawi 2013 female 45-49  positive  47564.08  40928.773  32704.094
#> 125 Malawi 2014 female 45-49  positive  51669.07  45922.710  38462.756
#> 126 Malawi 2015 female 45-49  positive  55224.80  50262.822  43736.400
#> 127 Malawi 2016 female 45-49  positive  59658.80  55376.540  50099.865
#> 128 Malawi 2017 female 45-49  positive  63983.38  60677.980  56571.297
#> 129 Malawi 2018 female 45-49  positive  68156.74  65617.480  62480.501
#> 130 Malawi 2019 female 45-49  positive  72233.94  70199.431  67462.751
#> 131 Malawi 2010   male 45-49  positive  35346.00  22261.737  12116.191
#> 132 Malawi 2011   male 45-49  positive  37782.04  25907.204  15414.116
#> 133 Malawi 2012   male 45-49  positive  40475.09  29465.706  19172.112
#> 134 Malawi 2013   male 45-49  positive  43270.12  32860.053  22969.549
#> 135 Malawi 2014   male 45-49  positive  45875.95  36123.931  26306.807
#> 136 Malawi 2015   male 45-49  positive  47734.46  38706.559  29311.084
#> 137 Malawi 2016   male 45-49  positive  50161.26  41752.013  32915.497
#> 138 Malawi 2017   male 45-49  positive  52265.32  44922.080  37108.249
#> 139 Malawi 2018   male 45-49  positive  54055.63  47758.901  41060.915
#> 140 Malawi 2019   male 45-49  positive  55678.55  50175.225  43112.541
#> 141 Malawi 2010 female 50-99  positive  48585.35  31080.990  22228.317
#> 142 Malawi 2011 female 50-99  positive  52924.99  37966.680  28397.011
#> 143 Malawi 2012 female 50-99  positive  57954.27  45187.836  36520.955
#> 144 Malawi 2013 female 50-99  positive  63730.36  52375.421  45113.356
#> 145 Malawi 2014 female 50-99  positive  70249.14  60132.772  53523.942
#> 146 Malawi 2015 female 50-99  positive  77506.60  68453.978  62535.362
#> 147 Malawi 2016 female 50-99  positive  85495.13  77522.008  72937.440
#> 148 Malawi 2017 female 50-99  positive  94135.34  87711.164  84264.927
#> 149 Malawi 2018 female 50-99  positive 103330.70  98240.015  95657.811
#> 150 Malawi 2019 female 50-99  positive 112902.74 108760.259 106351.763
#> 151 Malawi 2010   male 50-99  positive  45512.88  26531.933  17394.878
#> 152 Malawi 2011   male 50-99  positive  49160.12  31773.506  22280.164
#> 153 Malawi 2012   male 50-99  positive  53335.06  37175.835  27928.195
#> 154 Malawi 2013   male 50-99  positive  58033.76  42771.942  33975.204
#> 155 Malawi 2014   male 50-99  positive  63193.90  48830.506  39856.594
#> 156 Malawi 2015   male 50-99  positive  68802.85  55255.555  46311.139
#> 157 Malawi 2016   male 50-99  positive  74846.17  62203.461  53667.589
#> 158 Malawi 2017   male 50-99  positive  81251.51  70055.339  62501.518
#> 159 Malawi 2018   male 50-99  positive  87904.32  78164.502  71649.545
#> 160 Malawi 2019   male 50-99  positive  94663.52  86075.591  78521.665