Get indicator metadata

Get indicator metadata

get_meta_indicator()

Value

data.frame of indicator ids, labels, descriptions, and parameter mapping.

Examples

get_meta_indicator()

#>                indicator              indicator_label
#> 1             population                   Population
#> 2             prevalence               HIV prevalence
#> 3                  plhiv                        PLHIV
#> 4           art_coverage                 ART coverage
#> 5  art_current_residents       ART number (residents)
#> 6            art_current       ART number (attending)
#> 7    untreated_plhiv_num             PLHIV not on ART
#> 8       aware_plhiv_prop       Proportion PLHIV aware
#> 9      unaware_plhiv_num         Number PLHIV unaware
#> 10       aware_plhiv_num           Number PLHIV aware
#> 11             incidence                HIV incidence
#> 12            infections               New infections
#> 13        anc_prevalence           ANC HIV prevalence
#> 14      anc_art_coverage       ANC prior ART coverage
#> 15           anc_clients                  ANC clients
#> 16             anc_plhiv   HIV positive ANC attendees
#> 17       anc_already_art ANC attendees already on ART
#> 18           anc_art_new       ART initiations at ANC
#> 19         anc_known_pos           ANC known positive
#> 20        anc_tested_pos          ANC tested positive
#> 21        anc_tested_neg          ANC tested negative
#>                                                                      description
#> 1                                                                Population size
#> 2                                    Proportion of total population HIV positive
#> 3                                               Number of people living with HIV
#> 4                                         Proportion of PLHIV on ART (residents)
#> 5                                                      Number on ART (residents)
#> 6                                               Number receiving ART (attending)
#> 7                                                     Number of PLHIV not on ART
#> 8                               Proportion of PLHIV aware of HIV positive status
#> 9                                    Number PLHIV unaware of HIV positive status
#> 10                                               INDICATOR_LABEL_AWARE_PLHIV_NUM
#> 11                                                   HIV incidence rate per year
#> 12                                             Number of new infections per year
#> 13                                            HIV prevalence among ANC attendees
#> 14                           ART coverage among ANC attendees prior to first ANC
#> 15                                              Total number attending first ANC
#> 16                                    Number of HIV positive women attending ANC
#> 17                          Number already on ART prior prior to first ANC visit
#> 18                              Number of pregnant women initiated on ART at ANC
#> 19 Number of HIV positive ANC attendees aware of status prior to first ANC visit
#> 20                            Number of HIV positive ANC attendees tested at ANC
#> 21                            Number of HIV negative ANC attendees tested at ANC
#>                  parameter indicator_sort_order anc_indicator format scale
#> 1           population_out                    1         FALSE     NA    NA
#> 2                  rho_out                    2         FALSE     NA    NA
#> 3                plhiv_out                    3         FALSE     NA    NA
#> 4                alpha_out                    4         FALSE     NA    NA
#> 5               artnum_out                    5         FALSE     NA    NA
#> 6            artattend_out                    6         FALSE     NA    NA
#> 7  untreated_plhiv_num_out                    7         FALSE     NA    NA
#> 8     aware_plhiv_prop_out                    8         FALSE     NA    NA
#> 9    unaware_plhiv_num_out                    9         FALSE     NA    NA
#> 10     aware_plhiv_num_out                   10         FALSE     NA    NA
#> 11              lambda_out                   11         FALSE     NA    NA
#> 12          infections_out                   12         FALSE     NA    NA
#> 13             anc_rho_out                   13          TRUE     NA    NA
#> 14           anc_alpha_out                   14          TRUE     NA    NA
#> 15         anc_clients_out                   15          TRUE     NA    NA
#> 16           anc_plhiv_out                   16          TRUE     NA    NA
#> 17     anc_already_art_out                   17          TRUE     NA    NA
#> 18         anc_art_new_out                   18          TRUE     NA    NA
#> 19       anc_known_pos_out                   19          TRUE     NA    NA
#> 20      anc_tested_pos_out                   20          TRUE     NA    NA
#> 21      anc_tested_neg_out                   21          TRUE     NA    NA