Get JSON template representing model run options

This reads JSON file of model run options which describe how the Naomi front end should display input options. Returns template for the API to enrich with model options.

get_model_options_template(art, anc)

Arguments

art
If FALSE then don't return template for ART control section.

anc
If FALSE then don't return template for ANC control section.

Value

Model run options template.

Examples

get_model_options_template(TRUE, TRUE)

#> $general
#> [1] "{  \"label\": \"General\",  \"description\": \"Select general model options:\",  \"controlGroups\": [    {      \"label\": \"Area scope\",      \"controls\": [        {          \"name\": \"area_scope\",          \"type\": \"multiselect\",          \"options\":            <+area_scope_options+>,          \"value\": <+area_scope_default+>,          \"required\": true        }      ]    },    {      \"label\": \"Area level\",      \"controls\": [        {          \"name\": \"area_level\",          \"type\": \"select\",          \"options\":            <+area_level_options+>,          \"value\": <+area_level_default+>,          \"required\": true        }      ]    },    {      \"label\": \"Calendar quarter to generate estimates\",      \"controls\": [        {          \"name\": \"calendar_quarter_t2\",          \"type\": \"select\",          \"options\":            <+calendar_quarter_t2_options+>,          \"value\": \"CY2020Q4\",          \"required\": true        }      ]    },    {      \"label\": \"Calendar quarter to generate short-term projection\",      \"controls\": [        {          \"name\": \"calendar_quarter_t3\",          \"type\": \"select\",          \"value\": \"CY2021Q3\",          \"helpText\": \"Calendar quarter to generate short-term projection. Must be after calendar quarter to generate estimates.\",          \"required\": true,          \"options\": [            {              \"id\": \"CY2020Q1\",              \"label\": \"March 2020\"            },            {              \"id\": \"CY2020Q2\",              \"label\": \"June 2020\"            },            {              \"id\": \"CY2020Q3\",              \"label\": \"September 2020\"            },            {              \"id\": \"CY2020Q4\",              \"label\": \"December 2020\"            },            {              \"id\": \"CY2021Q1\",              \"label\": \"March 2021\"            },            {              \"id\": \"CY2021Q2\",              \"label\": \"June 2021\"            },            {              \"id\": \"CY2021Q3\",              \"label\": \"September 2021\"            },            {              \"id\": \"CY2021Q4\",              \"label\": \"December 2021\"            },            {              \"id\": \"CY2022Q1\",              \"label\": \"March 2022\"            },            {              \"id\": \"CY2022Q2\",              \"label\": \"June 2022\"            },            {              \"id\": \"CY2022Q3\",              \"label\": \"September 2022\"            },            {              \"id\": \"CY2022Q4\",              \"label\": \"December 2022\"            }          ]        }      ]    }  ]}"
#> 
#> $survey
#> [1] "{  \"label\": \"Survey\",  \"description\": \"Select which survey to use for each indicator and whether to run the model with survey ART coverage or VLS.\",  \"controlGroups\": [    {      \"label\": \"Calendar quarter at midpoint of survey\",      \"controls\": [        {          \"name\": \"calendar_quarter_t1\",          \"type\": \"select\",          \"options\":            <+calendar_quarter_t1_options+>,          \"value\": <+calendar_quarter_t1_default+>,          \"required\": true        }      ]    },    {      \"label\": \"Prevalence\",      \"controls\": [        {          \"name\": \"survey_prevalence\",          \"type\": \"multiselect\",          \"options\":            <+survey_prevalence_options+>,          \"value\": <+survey_prevalence_default+>,          \"required\": true        }      ]    },    {      \"label\": \"ART coverage\",      \"controls\": [        {          \"name\": \"survey_art_coverage\",          \"type\": \"multiselect\",          \"options\":            <+survey_art_coverage_options+>,          \"value\": <+survey_art_coverage_default+>,          \"required\": false        }      ]    }  ]}"
#> 
#> $anc
#> [1] "{  \"label\": \"ANC\",  \"description\": \"Optionally select which years of data to use at time point 1 and 2\",  \"controlGroups\": [    {      \"label\": \"Data year to calibrate number of ANC clients\",      \"controls\": [        {          \"name\": \"anc_clients_year2\",          \"label\": \"Time 2\",          \"type\": \"select\",          \"helpText\": \"Year to generate estimates for\",          \"options\":            <+anc_prevalence_year2_options+>,          \"value\": <+anc_prevalence_year2_default+>,          \"required\": false        }      ]    },    {      \"label\": \"Months included in ANC reporting for ANC clients\",      \"controls\": [        {          \"name\": \"anc_clients_year2_num_months\",          \"type\": \"select\",          \"value\": \"12\",          \"helpText\": \"Number of months included in ANC reporting in this year.\",          \"required\": true,          \"options\": [            {                \"id\": \"12\",                \"label\": \"12 months\"            },            {                \"id\": \"11\",                \"label\": \"11 months\"            },            {                \"id\": \"10\",                \"label\": \"10 months\"            },            {                \"id\": \"9\",                \"label\": \"9 months\"            },            {                \"id\": \"8\",                \"label\": \"8 months\"            },            {                \"id\": \"7\",                \"label\": \"7 months\"            },            {                \"id\": \"6\",                \"label\": \"6 months\"            },            {                \"id\": \"5\",                \"label\": \"5 months\"            },            {                \"id\": \"4\",                \"label\": \"4 months\"            },            {                \"id\": \"3\",                \"label\": \"3 months\"            },            {                \"id\": \"2\",                \"label\": \"2 months\"            },            {                \"id\": \"1\",                \"label\": \"1 month\"            }          ]        }      ]    },    {      \"label\": \"Prevalence among ANC attendees\",      \"controls\": [        {          \"name\": \"anc_prevalence_year1\",          \"label\": \"Time 1\",          \"type\": \"select\",          \"helpText\": \"Year matching midpoint of survey\",          \"options\":            <+anc_prevalence_year1_options+>,          \"value\": <+anc_prevalence_year1_default+>,          \"required\": false        },        {          \"name\": \"anc_prevalence_year2\",          \"label\": \"Time 2\",          \"type\": \"select\",          \"helpText\": \"Year to generate estimates for\",          \"options\":            <+anc_prevalence_year2_options+>,          \"value\": <+anc_prevalence_year2_default+>,          \"required\": false        }      ]    },    {      \"label\": \"ART coverage prior to first ANC\",      \"controls\": [        {          \"name\": \"anc_art_coverage_year1\",          \"label\": \"Time 1\",          \"type\": \"select\",          \"helpText\": \"Year matching midpoint of survey\",          \"options\":            <+anc_art_coverage_year1_options+>,          \"value\": <+anc_art_coverage_year1_default+>,          \"required\": false        },        {          \"name\": \"anc_art_coverage_year2\",          \"label\": \"Time 2\",          \"type\": \"select\",          \"helpText\": \"Year to generate estimates for\",          \"options\":            <+anc_art_coverage_year2_options+>,          \"value\": <+anc_art_coverage_year2_default+>,          \"required\": false        }      ]    }  ]}"
#> 
#> $art
#> [1] "{  \"label\": \"ART\",  \"description\": \"Select whether to include ART data or not in model run\",  \"controlGroups\": [    {      \"label\": \"Include ART data\",      \"controls\": [        {          \"name\": \"include_art_t1\",          \"label\": \"Time 1\",          \"type\": \"select\",          \"helpText\": \"Year matching midpoint of survey\",          \"options\": [            {              \"id\": \"true\",              \"label\": \"Yes\"            },            {              \"id\": \"false\",              \"label\": \"No\"            }          ],          \"required\": true        },        {          \"name\": \"include_art_t2\",          \"label\": \"Time 2\",          \"type\": \"select\",          \"helpText\": \"Year to generate estimates for\",          \"options\": [            {              \"id\": \"true\",              \"label\": \"Yes\"            },            {              \"id\": \"false\",              \"label\": \"No\"            }          ],          \"required\": true        }      ]    },    {      \"label\": \"Estimate neighbouring district ART attendance\",      \"controls\": [        {          \"name\": \"artattend\",          \"type\": \"select\",          \"required\": true,          \"options\": [            {              \"id\": \"false\",              \"label\": \"No\"            },            {              \"id\": \"true\",              \"label\": \"Yes\"            }          ],          \"value\": \"true\",          \"helpText\": \"Model proportion of ART residents who seek treatment in neighboring districts\"        }      ]    },    {      \"label\": \"Time-varying neighbouring district ART attendance\",      \"controls\": [        {          \"name\": \"artattend_t2\",          \"type\": \"select\",          \"required\": true,          \"options\": [            {              \"id\": \"false\",              \"label\": \"No\"            },            {              \"id\": \"true\",              \"label\": \"Yes\"            }          ],          \"value\": \"true\",          \"helpText\": \"Allow the odds for seeking treatment in neighbouring district to change over time.\"        }      ]    }  ]}"
#> 
#> $advanced
#> [1] "{  \"label\": \"Population calibration\",  \"description\": \"Options for calibrating to Spectrum population size. If national Spectrum file, 'National' and 'Subnational' options will return the same result. Calibration to Spectrum population is completed before model fitting.\",  \"controlGroups\": [    {      \"label\": \"Adjust to Spectrum population size\",      \"controls\": [        {          \"name\": \"spectrum_population_calibration\",          \"type\": \"select\",          \"options\": [            {              \"id\": \"none\",              \"label\": \"None\"            },            {              \"id\": \"national\",              \"label\": \"National\"            },            {              \"id\": \"subnational\",              \"label\": \"Subnational\"            }          ],          \"value\": \"national\",          \"required\": true,          \"helpText\": \"Calibrate to Spectrum population size by sex and 5-year age group by national or subnational. If national Spectrum file, national and subnational will return the same result.\"        }      ]    }  ]},{  \"label\": \"Advanced\",  \"description\": \"Advanced model run options\",  \"collapsible\": true,  \"collapsed\": true,  \"controlGroups\": [    {      \"label\": \"Output 'awareness of HIV status' indicators\",      \"controls\": [        {          \"name\": \"output_aware_plhiv\",          \"type\": \"select\",          \"value\": \"true\",          \"helpText\": \"Calculating awareness of HIV status indicators requires .shiny90 files to be included in each PJNZ. If .shiny90 files are not available, select 'No' to proceed with model fitting.\",          \"options\": [            {              \"id\": \"true\",              \"label\": \"Yes\"            },            {              \"id\": \"false\",              \"label\": \"No\"            }          ],          \"required\": true        }      ]    },    {      \"label\": \"Maximum iterations\",      \"controls\": [        {          \"name\": \"max_iterations\",          \"type\": \"number\",          \"value\": 250,          \"helpText\": \"Maximum number of iteration for model run\",          \"required\": true        }      ]    },    {      \"label\": \"Number of simulations\",      \"controls\": [        {          \"name\": \"no_of_samples\",          \"type\": \"number\",          \"value\": 1000,          \"helpText\": \"Number of simulatons for generating uncertainty ranges\",          \"required\": true        }      ]    },    {      \"label\": \"Simulation seed\",      \"controls\": [        {          \"name\": \"rng_seed\",          \"type\": \"number\",          \"value\": 28,          \"required\": false,          \"helpText\": \"Random number generator seed for uncertainty sample\"        }      ]    },    {      \"label\": \"Permissive\",      \"controls\": [        {          \"name\": \"permissive\",          \"type\": \"select\",          \"required\": true,          \"options\": [            {              \"id\": \"true\",              \"label\": \"Yes\"            },            {              \"id\": \"false\",              \"label\": \"No\"            }          ],          \"value\": \"false\",          \"helpText\": \"Select 'Yes' to allow fits with convergence or data errors\"        }      ]    },    {      \"label\": \"Logit prior mean for neighboring district attraction\",      \"controls\": [        {          \"name\": \"artattend_log_gamma_offset\",          \"type\": \"number\",          \"required\": true,          \"value\": -4,          \"helpText\": \"Logit offset for odds of obtaining ART in neighbouring district compared to home district\"        }      ]    },    {      \"label\": \"Use aggregate survey dataset\",      \"controls\": [        {          \"name\": \"use_survey_aggregate\",          \"type\": \"select\",          \"required\": true,          \"options\": [            {              \"id\": \"true\",              \"label\": \"Yes\"            },            {              \"id\": \"false\",              \"label\": \"No\"            }          ],          \"value\": \"false\",          \"helpText\": \"The uploaded survey dataset must be subsetted to exactly the data to be used in model calibration.\"        }      ]    },    {      \"label\": \"Use survey data on proportion recently infected\",      \"controls\": [        {          \"name\": \"survey_recently_infected\",          \"type\": \"multiselect\",          \"options\":            <+survey_recently_infected_options+>,          \"required\": false        }      ]    },          {      \"label\": \"Survey design effect - HIV prevalence\",      \"controls\": [        {          \"name\": \"use_kish_prev\",          \"label\": \"Use Kish ESS\",          \"type\": \"select\",          \"required\": true,          \"options\": [            {              \"id\": \"true\",              \"label\": \"Yes\"            },            {              \"id\": \"false\",              \"label\": \"No\"            }          ],          \"value\": \"true\",          \"helpText\": \"Use Kish Effective Sample Size (ESS) for survey data likelihood\"        },        {          \"name\": \"deff_prev\",          \"label\": \"ESS Scaling Factor\",          \"type\": \"number\",          \"required\": true,          \"value\": 1.0,          \"helpText\": \"Approximate DEFF to scale effective sample size for survey for HIV prevalence observations\"        }      ]    },    {      \"label\": \"Survey design effect - ART coverage\",      \"controls\": [        {          \"name\": \"use_kish_artcov\",          \"label\": \"Use Kish ESS\",          \"type\": \"select\",          \"required\": true,          \"options\": [            {              \"id\": \"true\",              \"label\": \"Yes\"            },            {              \"id\": \"false\",              \"label\": \"No\"            }          ],          \"value\": \"true\",          \"helpText\": \"Use Kish Effective Sample Size (ESS) for survey data likelihood\"        },        {          \"name\": \"deff_artcov\",          \"label\": \"ESS Scaling Factor\",          \"type\": \"number\",          \"required\": true,          \"value\": 1.0,          \"helpText\": \"Approximate DEFF to scale effective sample size for survey for ART coverage\"        }      ]    },    {      \"label\": \"Survey design effect - Proportion recently infected\",      \"controls\": [        {          \"name\": \"use_kish_recent\",          \"label\": \"Use Kish ESS\",          \"type\": \"select\",          \"required\": true,          \"options\": [            {              \"id\": \"true\",              \"label\": \"Yes\"            },            {              \"id\": \"false\",              \"label\": \"No\"            }          ],          \"value\": \"true\",          \"helpText\": \"Use Kish Effective Sample Size (ESS) for survey data likelihood\"        },        {          \"name\": \"deff_recent\",          \"label\": \"ESS Scaling Factor\",          \"type\": \"number\",          \"required\": true,          \"value\": 1.0,          \"helpText\": \"Approximate DEFF to scale effective sample size for survey for proportion recently infected amongst HIV positive\"        }      ]    }  ]}"
#> 

get_model_options_template(FALSE, FALSE)

#> $general
#> [1] "{  \"label\": \"General\",  \"description\": \"Select general model options:\",  \"controlGroups\": [    {      \"label\": \"Area scope\",      \"controls\": [        {          \"name\": \"area_scope\",          \"type\": \"multiselect\",          \"options\":            <+area_scope_options+>,          \"value\": <+area_scope_default+>,          \"required\": true        }      ]    },    {      \"label\": \"Area level\",      \"controls\": [        {          \"name\": \"area_level\",          \"type\": \"select\",          \"options\":            <+area_level_options+>,          \"value\": <+area_level_default+>,          \"required\": true        }      ]    },    {      \"label\": \"Calendar quarter to generate estimates\",      \"controls\": [        {          \"name\": \"calendar_quarter_t2\",          \"type\": \"select\",          \"options\":            <+calendar_quarter_t2_options+>,          \"value\": \"CY2020Q4\",          \"required\": true        }      ]    },    {      \"label\": \"Calendar quarter to generate short-term projection\",      \"controls\": [        {          \"name\": \"calendar_quarter_t3\",          \"type\": \"select\",          \"value\": \"CY2021Q3\",          \"helpText\": \"Calendar quarter to generate short-term projection. Must be after calendar quarter to generate estimates.\",          \"required\": true,          \"options\": [            {              \"id\": \"CY2020Q1\",              \"label\": \"March 2020\"            },            {              \"id\": \"CY2020Q2\",              \"label\": \"June 2020\"            },            {              \"id\": \"CY2020Q3\",              \"label\": \"September 2020\"            },            {              \"id\": \"CY2020Q4\",              \"label\": \"December 2020\"            },            {              \"id\": \"CY2021Q1\",              \"label\": \"March 2021\"            },            {              \"id\": \"CY2021Q2\",              \"label\": \"June 2021\"            },            {              \"id\": \"CY2021Q3\",              \"label\": \"September 2021\"            },            {              \"id\": \"CY2021Q4\",              \"label\": \"December 2021\"            },            {              \"id\": \"CY2022Q1\",              \"label\": \"March 2022\"            },            {              \"id\": \"CY2022Q2\",              \"label\": \"June 2022\"            },            {              \"id\": \"CY2022Q3\",              \"label\": \"September 2022\"            },            {              \"id\": \"CY2022Q4\",              \"label\": \"December 2022\"            }          ]        }      ]    }  ]}"
#> 
#> $survey
#> [1] "{  \"label\": \"Survey\",  \"description\": \"Select which survey to use for each indicator and whether to run the model with survey ART coverage or VLS.\",  \"controlGroups\": [    {      \"label\": \"Calendar quarter at midpoint of survey\",      \"controls\": [        {          \"name\": \"calendar_quarter_t1\",          \"type\": \"select\",          \"options\":            <+calendar_quarter_t1_options+>,          \"value\": <+calendar_quarter_t1_default+>,          \"required\": true        }      ]    },    {      \"label\": \"Prevalence\",      \"controls\": [        {          \"name\": \"survey_prevalence\",          \"type\": \"multiselect\",          \"options\":            <+survey_prevalence_options+>,          \"value\": <+survey_prevalence_default+>,          \"required\": true        }      ]    },    {      \"label\": \"ART coverage\",      \"controls\": [        {          \"name\": \"survey_art_coverage\",          \"type\": \"multiselect\",          \"options\":            <+survey_art_coverage_options+>,          \"value\": <+survey_art_coverage_default+>,          \"required\": false        }      ]    }  ]}"
#> 
#> $advanced
#> [1] "{  \"label\": \"Population calibration\",  \"description\": \"Options for calibrating to Spectrum population size. If national Spectrum file, 'National' and 'Subnational' options will return the same result. Calibration to Spectrum population is completed before model fitting.\",  \"controlGroups\": [    {      \"label\": \"Adjust to Spectrum population size\",      \"controls\": [        {          \"name\": \"spectrum_population_calibration\",          \"type\": \"select\",          \"options\": [            {              \"id\": \"none\",              \"label\": \"None\"            },            {              \"id\": \"national\",              \"label\": \"National\"            },            {              \"id\": \"subnational\",              \"label\": \"Subnational\"            }          ],          \"value\": \"national\",          \"required\": true,          \"helpText\": \"Calibrate to Spectrum population size by sex and 5-year age group by national or subnational. If national Spectrum file, national and subnational will return the same result.\"        }      ]    }  ]},{  \"label\": \"Advanced\",  \"description\": \"Advanced model run options\",  \"collapsible\": true,  \"collapsed\": true,  \"controlGroups\": [    {      \"label\": \"Output 'awareness of HIV status' indicators\",      \"controls\": [        {          \"name\": \"output_aware_plhiv\",          \"type\": \"select\",          \"value\": \"true\",          \"helpText\": \"Calculating awareness of HIV status indicators requires .shiny90 files to be included in each PJNZ. If .shiny90 files are not available, select 'No' to proceed with model fitting.\",          \"options\": [            {              \"id\": \"true\",              \"label\": \"Yes\"            },            {              \"id\": \"false\",              \"label\": \"No\"            }          ],          \"required\": true        }      ]    },    {      \"label\": \"Maximum iterations\",      \"controls\": [        {          \"name\": \"max_iterations\",          \"type\": \"number\",          \"value\": 250,          \"helpText\": \"Maximum number of iteration for model run\",          \"required\": true        }      ]    },    {      \"label\": \"Number of simulations\",      \"controls\": [        {          \"name\": \"no_of_samples\",          \"type\": \"number\",          \"value\": 1000,          \"helpText\": \"Number of simulatons for generating uncertainty ranges\",          \"required\": true        }      ]    },    {      \"label\": \"Simulation seed\",      \"controls\": [        {          \"name\": \"rng_seed\",          \"type\": \"number\",          \"value\": 28,          \"required\": false,          \"helpText\": \"Random number generator seed for uncertainty sample\"        }      ]    },    {      \"label\": \"Permissive\",      \"controls\": [        {          \"name\": \"permissive\",          \"type\": \"select\",          \"required\": true,          \"options\": [            {              \"id\": \"true\",              \"label\": \"Yes\"            },            {              \"id\": \"false\",              \"label\": \"No\"            }          ],          \"value\": \"false\",          \"helpText\": \"Select 'Yes' to allow fits with convergence or data errors\"        }      ]    },    {      \"label\": \"Logit prior mean for neighboring district attraction\",      \"controls\": [        {          \"name\": \"artattend_log_gamma_offset\",          \"type\": \"number\",          \"required\": true,          \"value\": -4,          \"helpText\": \"Logit offset for odds of obtaining ART in neighbouring district compared to home district\"        }      ]    },    {      \"label\": \"Use aggregate survey dataset\",      \"controls\": [        {          \"name\": \"use_survey_aggregate\",          \"type\": \"select\",          \"required\": true,          \"options\": [            {              \"id\": \"true\",              \"label\": \"Yes\"            },            {              \"id\": \"false\",              \"label\": \"No\"            }          ],          \"value\": \"false\",          \"helpText\": \"The uploaded survey dataset must be subsetted to exactly the data to be used in model calibration.\"        }      ]    },    {      \"label\": \"Use survey data on proportion recently infected\",      \"controls\": [        {          \"name\": \"survey_recently_infected\",          \"type\": \"multiselect\",          \"options\":            <+survey_recently_infected_options+>,          \"required\": false        }      ]    },          {      \"label\": \"Survey design effect - HIV prevalence\",      \"controls\": [        {          \"name\": \"use_kish_prev\",          \"label\": \"Use Kish ESS\",          \"type\": \"select\",          \"required\": true,          \"options\": [            {              \"id\": \"true\",              \"label\": \"Yes\"            },            {              \"id\": \"false\",              \"label\": \"No\"            }          ],          \"value\": \"true\",          \"helpText\": \"Use Kish Effective Sample Size (ESS) for survey data likelihood\"        },        {          \"name\": \"deff_prev\",          \"label\": \"ESS Scaling Factor\",          \"type\": \"number\",          \"required\": true,          \"value\": 1.0,          \"helpText\": \"Approximate DEFF to scale effective sample size for survey for HIV prevalence observations\"        }      ]    },    {      \"label\": \"Survey design effect - ART coverage\",      \"controls\": [        {          \"name\": \"use_kish_artcov\",          \"label\": \"Use Kish ESS\",          \"type\": \"select\",          \"required\": true,          \"options\": [            {              \"id\": \"true\",              \"label\": \"Yes\"            },            {              \"id\": \"false\",              \"label\": \"No\"            }          ],          \"value\": \"true\",          \"helpText\": \"Use Kish Effective Sample Size (ESS) for survey data likelihood\"        },        {          \"name\": \"deff_artcov\",          \"label\": \"ESS Scaling Factor\",          \"type\": \"number\",          \"required\": true,          \"value\": 1.0,          \"helpText\": \"Approximate DEFF to scale effective sample size for survey for ART coverage\"        }      ]    },    {      \"label\": \"Survey design effect - Proportion recently infected\",      \"controls\": [        {          \"name\": \"use_kish_recent\",          \"label\": \"Use Kish ESS\",          \"type\": \"select\",          \"required\": true,          \"options\": [            {              \"id\": \"true\",              \"label\": \"Yes\"            },            {              \"id\": \"false\",              \"label\": \"No\"            }          ],          \"value\": \"true\",          \"helpText\": \"Use Kish Effective Sample Size (ESS) for survey data likelihood\"        },        {          \"name\": \"deff_recent\",          \"label\": \"ESS Scaling Factor\",          \"type\": \"number\",          \"required\": true,          \"value\": 1.0,          \"helpText\": \"Approximate DEFF to scale effective sample size for survey for proportion recently infected amongst HIV positive\"        }      ]    }  ]}"
#>