Get plotting metadata for a particular country

Get plotting metadata for a particular country

get_plotting_metadata(iso3)

Arguments

iso3

iso3 code of country to get metadata for or blank for default configuration.

Value

List of plotting metadata about how to locate data for a specific indicator, data type and plot type. Also returns metadata about colour scheme to use for that country and indicator.

Examples

get_plotting_metadata("MWI")
#>                        indicator        data_type  plot_type value_column
#> 1                    art_current        programme choropleth        value
#> 2                    art_current        calibrate   barchart         mean
#> 3                    art_current           output   barchart         mean
#> 4                    art_current       comparison   barchart         mean
#> 5                    art_current           output choropleth         mean
#> 6                    art_adult_f        programme choropleth        value
#> 7                    art_adult_m        programme choropleth        value
#> 8                      art_adult        programme choropleth        value
#> 9                      art_child        programme choropleth        value
#> 10          art_current_adjusted        programme choropleth        value
#> 11          art_adjusted_adult_f        programme choropleth        value
#> 12          art_adjusted_adult_m        programme choropleth        value
#> 13            art_adjusted_adult        programme choropleth        value
#> 14            art_adjusted_child        programme choropleth        value
#> 15           art_adult_sex_ratio        programme choropleth        value
#> 16         art_child_adult_ratio        programme choropleth        value
#> 17                anc_prevalence           output   barchart         mean
#> 18                anc_prevalence              anc choropleth        value
#> 19                anc_prevalence           output choropleth         mean
#> 20              anc_art_coverage              anc choropleth        value
#> 21              anc_art_coverage           output choropleth         mean
#> 22              anc_art_coverage           output   barchart         mean
#> 23                   anc_clients           output   barchart         mean
#> 24                   anc_clients           output choropleth         mean
#> 25                   anc_clients input_comparison   barchart        value
#> 26                   anc_clients              anc choropleth        value
#> 27                    anc_tested input_comparison   barchart        value
#> 28                    anc_tested              anc choropleth        value
#> 29                anc_tested_pos           output choropleth         mean
#> 30                anc_tested_pos           output   barchart         mean
#> 31                anc_tested_pos input_comparison   barchart        value
#> 32                anc_tested_pos              anc choropleth        value
#> 33                 anc_known_pos              anc choropleth        value
#> 34                 anc_known_pos           output   barchart         mean
#> 35                 anc_known_pos           output choropleth         mean
#> 36                 anc_known_pos input_comparison   barchart        value
#> 37                 anc_known_neg              anc choropleth        value
#> 38               anc_already_art           output   barchart         mean
#> 39               anc_already_art           output choropleth         mean
#> 40               anc_already_art input_comparison   barchart        value
#> 41               anc_already_art              anc choropleth        value
#> 42                 anc_total_pos              anc choropleth        value
#> 43                    anc_status              anc choropleth        value
#> 44           anc_art_among_known              anc choropleth        value
#> 45          births_clients_ratio              anc choropleth        value
#> 46               births_facility              anc choropleth        value
#> 47                 art_new_total        programme choropleth        value
#> 48               art_new_adult_f        programme choropleth        value
#> 49               art_new_adult_m        programme choropleth        value
#> 50                 art_new_adult        programme choropleth        value
#> 51                 art_new_child        programme choropleth        value
#> 52         vl_tested_12mos_total        programme choropleth        value
#> 53       vl_tested_12mos_adult_f        programme choropleth        value
#> 54       vl_tested_12mos_adult_m        programme choropleth        value
#> 55         vl_tested_12mos_adult        programme choropleth        value
#> 56         vl_tested_12mos_child        programme choropleth        value
#> 57     vl_suppressed_12mos_total        programme choropleth        value
#> 58   vl_suppressed_12mos_adult_f        programme choropleth        value
#> 59   vl_suppressed_12mos_adult_m        programme choropleth        value
#> 60     vl_suppressed_12mos_adult        programme choropleth        value
#> 61     vl_suppressed_12mos_child        programme choropleth        value
#> 62      vl_prop_suppressed_total        programme choropleth        value
#> 63    vl_prop_suppressed_adult_f        programme choropleth        value
#> 64    vl_prop_suppressed_adult_m        programme choropleth        value
#> 65      vl_prop_suppressed_adult        programme choropleth        value
#> 66      vl_prop_suppressed_child        programme choropleth        value
#> 67             vl_coverage_total        programme choropleth        value
#> 68           vl_coverage_adult_f        programme choropleth        value
#> 69           vl_coverage_adult_m        programme choropleth        value
#> 70             vl_coverage_adult        programme choropleth        value
#> 71             vl_coverage_child        programme choropleth        value
#> 72                    prevalence           output   barchart         mean
#> 73                    prevalence           survey choropleth     estimate
#> 74                    prevalence        calibrate   barchart         mean
#> 75                    prevalence           output choropleth         mean
#> 76                    prevalence       comparison   barchart         mean
#> 77                  art_coverage           survey choropleth     estimate
#> 78                  art_coverage           output   barchart         mean
#> 79                  art_coverage        calibrate   barchart         mean
#> 80                  art_coverage       comparison   barchart         mean
#> 81                  art_coverage           output choropleth         mean
#> 82               recent_infected           survey choropleth     estimate
#> 83       viral_suppression_plhiv           survey choropleth     estimate
#> 84         art_current_residents           output choropleth         mean
#> 85         art_current_residents           output   barchart         mean
#> 86                    population        calibrate   barchart         mean
#> 87                    population       population    pyramid   population
#> 88                    population           output choropleth         mean
#> 89                    population           output   barchart         mean
#> 90                         plhiv           output choropleth         mean
#> 91                         plhiv           output   barchart         mean
#> 92                         plhiv        calibrate   barchart         mean
#> 93                     incidence           output choropleth         mean
#> 94                     incidence        calibrate   barchart         mean
#> 95                     incidence           output   barchart         mean
#> 96                    infections           output choropleth         mean
#> 97                    infections           output   barchart         mean
#> 98                    infections        calibrate   barchart         mean
#> 99              aware_plhiv_prop           output choropleth         mean
#> 100             aware_plhiv_prop           output   barchart         mean
#> 101             aware_plhiv_prop        calibrate   barchart         mean
#> 102            unaware_plhiv_num        calibrate   barchart         mean
#> 103            unaware_plhiv_num           output choropleth         mean
#> 104            unaware_plhiv_num           output   barchart         mean
#> 105              aware_plhiv_num           output choropleth         mean
#> 106              aware_plhiv_num           output   barchart         mean
#> 107          untreated_plhiv_num           output   barchart         mean
#> 108          untreated_plhiv_num           output choropleth         mean
#> 109                 plhiv_attend           output   barchart         mean
#> 110                 plhiv_attend           output choropleth         mean
#> 111       untreated_plhiv_attend           output choropleth         mean
#> 112       untreated_plhiv_attend           output   barchart         mean
#> 113           aware_plhiv_attend           output choropleth         mean
#> 114           aware_plhiv_attend           output   barchart         mean
#> 115         unaware_plhiv_attend           output choropleth         mean
#> 116         unaware_plhiv_attend           output   barchart         mean
#> 117                    anc_plhiv           output choropleth         mean
#> 118                    anc_plhiv           output   barchart         mean
#> 119                  anc_art_new           output choropleth         mean
#> 120                  anc_art_new           output   barchart         mean
#> 121               anc_tested_neg           output choropleth         mean
#> 122               anc_tested_neg           output   barchart         mean
#> 123             population_ratio        calibrate   barchart         mean
#> 124                  plhiv_ratio        calibrate   barchart         mean
#> 125             infections_ratio        calibrate   barchart         mean
#> 126            art_current_ratio        calibrate   barchart         mean
#> 127      unaware_plhiv_num_ratio        calibrate   barchart         mean
#> 128             prevalence_ratio        calibrate   barchart         mean
#> 129           art_coverage_ratio        calibrate   barchart         mean
#> 130       aware_plhiv_prop_ratio        calibrate   barchart         mean
#> 131              incidence_ratio        calibrate   barchart         mean
#> 132   anc_prevalence_age_matched       comparison   barchart         mean
#> 133 anc_art_coverage_age_matched       comparison   barchart         mean
#> 134                number_on_art input_comparison   barchart        value
#> 135        population_proportion       population    pyramid   population
#>     error_low_column error_high_column indicator_column
#> 1                                                  plot
#> 2              lower             upper        indicator
#> 3              lower             upper        indicator
#> 4              lower             upper        indicator
#> 5              lower             upper        indicator
#> 6                                                  plot
#> 7                                                  plot
#> 8                                                  plot
#> 9                                                  plot
#> 10                                                 plot
#> 11                                                 plot
#> 12                                                 plot
#> 13                                                 plot
#> 14                                                 plot
#> 15                                                 plot
#> 16                                                 plot
#> 17             lower             upper        indicator
#> 18                                                 plot
#> 19             lower             upper        indicator
#> 20                                                 plot
#> 21             lower             upper        indicator
#> 22             lower             upper        indicator
#> 23             lower             upper        indicator
#> 24             lower             upper        indicator
#> 25                                                     
#> 26                                                 plot
#> 27                                                     
#> 28                                                 plot
#> 29             lower             upper        indicator
#> 30             lower             upper        indicator
#> 31                                                     
#> 32                                                 plot
#> 33                                                 plot
#> 34             lower             upper        indicator
#> 35             lower             upper        indicator
#> 36                                                     
#> 37                                                 plot
#> 38             lower             upper        indicator
#> 39             lower             upper        indicator
#> 40                                                     
#> 41                                                 plot
#> 42                                                 plot
#> 43                                                 plot
#> 44                                                 plot
#> 45                                                 plot
#> 46                                                 plot
#> 47                                                 plot
#> 48                                                 plot
#> 49                                                 plot
#> 50                                                 plot
#> 51                                                 plot
#> 52                                                 plot
#> 53                                                 plot
#> 54                                                 plot
#> 55                                                 plot
#> 56                                                 plot
#> 57                                                 plot
#> 58                                                 plot
#> 59                                                 plot
#> 60                                                 plot
#> 61                                                 plot
#> 62                                                 plot
#> 63                                                 plot
#> 64                                                 plot
#> 65                                                 plot
#> 66                                                 plot
#> 67                                                 plot
#> 68                                                 plot
#> 69                                                 plot
#> 70                                                 plot
#> 71                                                 plot
#> 72             lower             upper        indicator
#> 73          ci_lower          ci_upper        indicator
#> 74             lower             upper        indicator
#> 75             lower             upper        indicator
#> 76             lower             upper        indicator
#> 77          ci_lower          ci_upper        indicator
#> 78             lower             upper        indicator
#> 79             lower             upper        indicator
#> 80             lower             upper        indicator
#> 81             lower             upper        indicator
#> 82          ci_lower          ci_upper        indicator
#> 83          ci_lower          ci_upper        indicator
#> 84             lower             upper        indicator
#> 85             lower             upper        indicator
#> 86             lower             upper        indicator
#> 87                                                     
#> 88             lower             upper        indicator
#> 89             lower             upper        indicator
#> 90             lower             upper        indicator
#> 91             lower             upper        indicator
#> 92             lower             upper        indicator
#> 93             lower             upper        indicator
#> 94             lower             upper        indicator
#> 95             lower             upper        indicator
#> 96             lower             upper        indicator
#> 97             lower             upper        indicator
#> 98             lower             upper        indicator
#> 99             lower             upper        indicator
#> 100            lower             upper        indicator
#> 101            lower             upper        indicator
#> 102            lower             upper        indicator
#> 103            lower             upper        indicator
#> 104            lower             upper        indicator
#> 105            lower             upper        indicator
#> 106            lower             upper        indicator
#> 107            lower             upper        indicator
#> 108            lower             upper        indicator
#> 109            lower             upper        indicator
#> 110            lower             upper        indicator
#> 111            lower             upper        indicator
#> 112            lower             upper        indicator
#> 113            lower             upper        indicator
#> 114            lower             upper        indicator
#> 115            lower             upper        indicator
#> 116            lower             upper        indicator
#> 117            lower             upper        indicator
#> 118            lower             upper        indicator
#> 119            lower             upper        indicator
#> 120            lower             upper        indicator
#> 121            lower             upper        indicator
#> 122            lower             upper        indicator
#> 123            lower             upper        indicator
#> 124            lower             upper        indicator
#> 125            lower             upper        indicator
#> 126            lower             upper        indicator
#> 127            lower             upper        indicator
#> 128            lower             upper        indicator
#> 129            lower             upper        indicator
#> 130            lower             upper        indicator
#> 131            lower             upper        indicator
#> 132            lower             upper        indicator
#> 133            lower             upper        indicator
#> 134                                                    
#> 135                                                    
#>                  indicator_value                                  name scale
#> 1                    art_current                ART number (attending)     1
#> 2                    art_current                ART number (attending)     1
#> 3                    art_current                ART number (attending)     1
#> 4                    art_current                ART number (attending)     1
#> 5                    art_current                ART number (attending)     1
#> 6                    art_adult_f                            ART female     1
#> 7                    art_adult_m                              ART male     1
#> 8                      art_adult                             ART adult     1
#> 9                      art_child                        ART paediatric     1
#> 10          art_current_adjusted       ART number (residents) adjusted     1
#> 11          art_adjusted_adult_f                   ART female adjusted     1
#> 12          art_adjusted_adult_m                     ART male adjusted     1
#> 13            art_adjusted_adult                    ART adult adjusted     1
#> 14            art_adjusted_child               ART paediatric adjusted     1
#> 15           art_adult_sex_ratio                   ART adult sex ratio     1
#> 16         art_child_adult_ratio              ART child-to-adult ratio     1
#> 17                anc_prevalence                    ANC HIV prevalence     1
#> 18                anc_prevalence                    ANC HIV prevalence     1
#> 19                anc_prevalence                    ANC HIV prevalence     1
#> 20              anc_art_coverage                ANC prior ART coverage     1
#> 21              anc_art_coverage                ANC prior ART coverage     1
#> 22              anc_art_coverage                ANC prior ART coverage     1
#> 23                   anc_clients                           ANC clients     1
#> 24                   anc_clients                           ANC clients     1
#> 25                   anc_clients                           ANC clients     1
#> 26                   anc_clients                           ANC clients     1
#> 27                    anc_tested                            ANC tested     1
#> 28                    anc_tested                            ANC tested     1
#> 29                anc_tested_pos                   ANC tested positive     1
#> 30                anc_tested_pos                   ANC tested positive     1
#> 31                anc_tested_pos                   ANC tested positive     1
#> 32                anc_tested_pos                   ANC tested positive     1
#> 33                 anc_known_pos                    ANC known positive     1
#> 34                 anc_known_pos                    ANC known positive     1
#> 35                 anc_known_pos                    ANC known positive     1
#> 36                 anc_known_pos                    ANC known positive     1
#> 37                 anc_known_neg                    ANC known negative     1
#> 38               anc_already_art          ANC attendees already on ART     1
#> 39               anc_already_art          ANC attendees already on ART     1
#> 40               anc_already_art          ANC attendees already on ART     1
#> 41               anc_already_art          ANC attendees already on ART     1
#> 42                 anc_total_pos                ANC total HIV positive     1
#> 43                    anc_status            ANC HIV status ascertained     1
#> 44           anc_art_among_known ANC ART coverage among known positive     1
#> 45          births_clients_ratio                  Births-clients ratio     1
#> 46               births_facility                Health facility births     1
#> 47                 art_new_total                               ART new     1
#> 48               art_new_adult_f                        ART new female     1
#> 49               art_new_adult_m                          ART new male     1
#> 50                 art_new_adult                         ART new adult     1
#> 51                 art_new_child                    ART new paediatric     1
#> 52         vl_tested_12mos_total                             VL tested     1
#> 53       vl_tested_12mos_adult_f                      VL tested female     1
#> 54       vl_tested_12mos_adult_m                        VL tested male     1
#> 55         vl_tested_12mos_adult                       VL tested adult     1
#> 56         vl_tested_12mos_child                  VL tested paediatric     1
#> 57     vl_suppressed_12mos_total                   VL tests suppressed     1
#> 58   vl_suppressed_12mos_adult_f            VL tests suppressed female     1
#> 59   vl_suppressed_12mos_adult_m              VL tests suppressed male     1
#> 60     vl_suppressed_12mos_adult             VL tests suppressed adult     1
#> 61     vl_suppressed_12mos_child        VL tests suppressed paediatric     1
#> 62      vl_prop_suppressed_total                Viral load suppression     1
#> 63    vl_prop_suppressed_adult_f         Viral load suppression female     1
#> 64    vl_prop_suppressed_adult_m           Viral load suppression male     1
#> 65      vl_prop_suppressed_adult          Viral load suppression adult     1
#> 66      vl_prop_suppressed_child     Viral load suppression paediatric     1
#> 67             vl_coverage_total                   VL testing coverage     1
#> 68           vl_coverage_adult_f            VL testing coverage female     1
#> 69           vl_coverage_adult_m              VL testing coverage male     1
#> 70             vl_coverage_adult             VL testing coverage adult     1
#> 71             vl_coverage_child        VL testing coverage paediatric     1
#> 72                    prevalence                        HIV prevalence     1
#> 73                    prevalence                        HIV prevalence     1
#> 74                    prevalence                        HIV prevalence     1
#> 75                    prevalence                        HIV prevalence     1
#> 76                    prevalence                        HIV prevalence     1
#> 77                  art_coverage                          ART coverage     1
#> 78                  art_coverage                          ART coverage     1
#> 79                  art_coverage                          ART coverage     1
#> 80                  art_coverage                          ART coverage     1
#> 81                  art_coverage                          ART coverage     1
#> 82               recent_infected          Proportion recently infected     1
#> 83       viral_suppression_plhiv                Viral load suppression     1
#> 84         art_current_residents                ART number (residents)     1
#> 85         art_current_residents                ART number (residents)     1
#> 86                    population                            Population     1
#> 87                    population                            Population     1
#> 88                    population                            Population     1
#> 89                    population                            Population     1
#> 90                         plhiv                                 PLHIV     1
#> 91                         plhiv                                 PLHIV     1
#> 92                         plhiv                                 PLHIV     1
#> 93                     incidence                         HIV incidence  1000
#> 94                     incidence                         HIV incidence  1000
#> 95                     incidence                         HIV incidence  1000
#> 96                    infections                        New infections     1
#> 97                    infections                        New infections     1
#> 98                    infections                        New infections     1
#> 99              aware_plhiv_prop                Proportion PLHIV aware     1
#> 100             aware_plhiv_prop                Proportion PLHIV aware     1
#> 101             aware_plhiv_prop                Proportion PLHIV aware     1
#> 102            unaware_plhiv_num                  Number PLHIV unaware     1
#> 103            unaware_plhiv_num                  Number PLHIV unaware     1
#> 104            unaware_plhiv_num                  Number PLHIV unaware     1
#> 105              aware_plhiv_num                    Number PLHIV aware     1
#> 106              aware_plhiv_num                    Number PLHIV aware     1
#> 107          untreated_plhiv_num                      PLHIV not on ART     1
#> 108          untreated_plhiv_num                      PLHIV not on ART     1
#> 109                 plhiv_attend                 PLHIV (ART catchment)     1
#> 110                 plhiv_attend                 PLHIV (ART catchment)     1
#> 111       untreated_plhiv_attend       Untreated PLHIV (ART catchment)     1
#> 112       untreated_plhiv_attend       Untreated PLHIV (ART catchment)     1
#> 113           aware_plhiv_attend    Number Aware PLHIV (ART catchment)     1
#> 114           aware_plhiv_attend    Number Aware PLHIV (ART catchment)     1
#> 115         unaware_plhiv_attend  Number unaware PLHIV (ART catchment)     1
#> 116         unaware_plhiv_attend  Number unaware PLHIV (ART catchment)     1
#> 117                    anc_plhiv            HIV positive ANC attendees     1
#> 118                    anc_plhiv            HIV positive ANC attendees     1
#> 119                  anc_art_new                ART initiations at ANC     1
#> 120                  anc_art_new                ART initiations at ANC     1
#> 121               anc_tested_neg                   ANC tested negative     1
#> 122               anc_tested_neg                   ANC tested negative     1
#> 123             population_ratio                      Population ratio     1
#> 124                  plhiv_ratio                           PLHIV ratio     1
#> 125             infections_ratio                  New infections ratio     1
#> 126            art_current_ratio          ART number (attending) ratio     1
#> 127      unaware_plhiv_num_ratio            Number PLHIV unaware ratio     1
#> 128             prevalence_ratio                  HIV prevalence ratio     1
#> 129           art_coverage_ratio                    ART coverage ratio     1
#> 130       aware_plhiv_prop_ratio          Proportion PLHIV aware ratio     1
#> 131              incidence_ratio                       Incidence ratio     1
#> 132   anc_prevalence_age_matched            ANC prevalence age matched     1
#> 133 anc_art_coverage_age_matched          ANC ART coverage age matched     1
#> 134                number_on_art                         Number on ART     1
#> 135        population_proportion                 Population proportion     1
#>     accuracy format indicator_sort_order    iso3             colour min     max
#> 1         NA  "0,0"                    1     MWI  interpolateGreens   0 8.0e+04
#> 2        100    0,0                    6     MWI  interpolateGreens   0 8.0e+04
#> 3        100    0,0                    6     MWI  interpolateGreens   0 8.0e+04
#> 4        100    0,0                    6     MWI  interpolateGreens   0 8.0e+04
#> 5        100    0,0                    6     MWI  interpolateGreens   0 8.0e+04
#> 6         NA  "0,0"                    2 default  interpolateGreens   0 8.0e+04
#> 7         NA  "0,0"                    3 default  interpolateGreens   0 8.0e+04
#> 8         NA  "0,0"                    4 default  interpolateGreens   0 8.0e+04
#> 9         NA  "0,0"                    5 default  interpolateGreens   0 8.0e+04
#> 10        NA  "0,0"                    6 default  interpolateGreens   0 8.0e+04
#> 11        NA  "0,0"                    7 default  interpolateGreens   0 8.0e+04
#> 12        NA  "0,0"                    8 default  interpolateGreens   0 8.0e+04
#> 13        NA  "0,0"                    9 default  interpolateGreens   0 8.0e+04
#> 14        NA  "0,0"                   10 default  interpolateGreens   0 8.0e+04
#> 15        NA   "0%"                   11 default  interpolateGreens   0 1.2e+00
#> 16        NA   "0%"                   12 default  interpolateGreens   0 1.2e+00
#> 17        NA   0.0%                   18     MWI interpolateOranges   0 5.0e-01
#> 18        NA "0.0%"                   13     MWI interpolateOranges   0 5.0e-01
#> 19        NA   0.0%                   18     MWI interpolateOranges   0 5.0e-01
#> 20        NA   "0%"                   14     MWI interpolateOranges   0 1.0e+00
#> 21        NA   0.0%                   19     MWI interpolateOranges   0 1.0e+00
#> 22        NA   0.0%                   19     MWI interpolateOranges   0 1.0e+00
#> 23       100    0,0                   20 default interpolateOranges   0 7.0e+05
#> 24       100    0,0                   20 default interpolateOranges   0 7.0e+05
#> 25         1    0,0                   20 default interpolateOranges   0 7.0e+05
#> 26        NA  "0,0"                   15 default interpolateOranges   0 7.0e+05
#> 27         1    0,0                   25 default interpolateOranges   0 1.0e+06
#> 28        NA  "0,0"                   16 default interpolateOranges   0 1.0e+06
#> 29        10    0,0                   26 default interpolateOranges   0 3.5e+05
#> 30        10    0,0                   26 default interpolateOranges   0 3.5e+05
#> 31         1    0,0                   26 default interpolateOranges   0 3.5e+05
#> 32        NA  "0,0"                   17 default interpolateOranges   0 3.5e+05
#> 33        NA  "0,0"                   18 default interpolateOranges   0 3.5e+05
#> 34        10    0,0                   24 default interpolateOranges   0 3.5e+05
#> 35        10    0,0                   24 default interpolateOranges   0 3.5e+05
#> 36         1    0,0                   24 default interpolateOranges   0 3.5e+05
#> 37        NA  "0,0"                   19 default interpolateOranges   0 7.0e+05
#> 38        10    0,0                   22 default interpolateOranges   0 3.5e+05
#> 39        10    0,0                   22 default interpolateOranges   0 3.5e+05
#> 40         1    0,0                   22 default interpolateOranges   0 3.5e+05
#> 41        NA  "0,0"                   20 default interpolateOranges   0 3.5e+05
#> 42        NA  "0,0"                   21 default interpolateOranges   0 3.5e+05
#> 43        NA  "0,0"                   22 default interpolateOranges   0 3.5e+05
#> 44        NA   "0%"                   23 default interpolateOranges   0 3.5e+05
#> 45        NA   "0%"                   24 default   interpolateGreys   0 1.2e+00
#> 46        NA  "0,0"                   25 default   interpolateGreys   0 3.5e+05
#> 47        NA  "0,0"                   26 default  interpolateGreens   0 8.0e+04
#> 48        NA  "0,0"                   27 default  interpolateGreens   0 8.0e+04
#> 49        NA  "0,0"                   28 default  interpolateGreens   0 8.0e+04
#> 50        NA  "0,0"                   29 default  interpolateGreens   0 8.0e+04
#> 51        NA  "0,0"                   30 default  interpolateGreens   0 8.0e+04
#> 52        NA  "0,0"                   31 default interpolatePurples   0 6.0e+04
#> 53        NA  "0,0"                   32 default interpolatePurples   0 6.0e+04
#> 54        NA  "0,0"                   33 default interpolatePurples   0 6.0e+04
#> 55        NA  "0,0"                   34 default interpolatePurples   0 6.0e+04
#> 56        NA  "0,0"                   35 default interpolatePurples   0 6.0e+04
#> 57        NA  "0,0"                   36 default interpolatePurples   0 6.0e+04
#> 58        NA  "0,0"                   37 default interpolatePurples   0 6.0e+04
#> 59        NA  "0,0"                   38 default interpolatePurples   0 6.0e+04
#> 60        NA  "0,0"                   39 default interpolatePurples   0 6.0e+04
#> 61        NA  "0,0"                   40 default interpolatePurples   0 6.0e+04
#> 62        NA   "0%"                   41 default interpolatePurples   0 1.2e+00
#> 63        NA   "0%"                   42 default interpolatePurples   0 1.2e+00
#> 64        NA   "0%"                   43 default interpolatePurples   0 1.2e+00
#> 65        NA   "0%"                   44 default interpolatePurples   0 1.2e+00
#> 66        NA   "0%"                   45 default interpolatePurples   0 1.2e+00
#> 67        NA   "0%"                   46 default interpolatePurples   0 1.2e+00
#> 68        NA   "0%"                   47 default interpolatePurples   0 1.2e+00
#> 69        NA   "0%"                   48 default interpolatePurples   0 1.2e+00
#> 70        NA   "0%"                   49 default interpolatePurples   0 1.2e+00
#> 71        NA   "0%"                   50 default interpolatePurples   0 1.2e+00
#> 72        NA   0.0%                    2     MWI    interpolateReds   0 5.0e-01
#> 73        NA   0.0%                    2     MWI    interpolateReds   0 5.0e-01
#> 74        NA   0.0%                    2     MWI    interpolateReds   0 5.0e-01
#> 75        NA   0.0%                    2     MWI    interpolateReds   0 5.0e-01
#> 76        NA   0.0%                    2     MWI    interpolateReds   0 5.0e-01
#> 77        NA   0.0%                    4     MWI  interpolateGreens   0 1.0e+00
#> 78        NA   0.0%                    4     MWI  interpolateGreens   0 1.0e+00
#> 79        NA   0.0%                    4     MWI  interpolateGreens   0 1.0e+00
#> 80        NA   0.0%                    4     MWI  interpolateGreens   0 1.0e+00
#> 81        NA   0.0%                    4     MWI  interpolateGreens   0 1.0e+00
#> 82         1  0.00%                   NA     MWI   interpolateBlues   0 6.0e-01
#> 83        NA   0.0%                   NA     MWI interpolatePurples   0 1.0e+00
#> 84       100    0,0                    5     MWI  interpolateGreens   0 7.0e+04
#> 85       100    0,0                    5     MWI  interpolateGreens   0 7.0e+04
#> 86      1000    0,0                    1     MWI   interpolateGreys   0 2.0e+07
#> 87         1    0,0                    1     MWI   interpolateGreys   0 2.0e+07
#> 88      1000    0,0                    1     MWI   interpolateGreys   0 2.0e+07
#> 89      1000    0,0                    1     MWI   interpolateGreys   0 2.0e+07
#> 90       100    0,0                    3     MWI    interpolateReds   0 1.0e+07
#> 91       100    0,0                    3     MWI    interpolateReds   0 1.0e+07
#> 92       100    0,0                    3     MWI    interpolateReds   0 1.0e+07
#> 93        NA    0.0                   16     MWI   interpolateBlues   0 1.0e-02
#> 94        NA    0.0                   16     MWI   interpolateBlues   0 1.0e-02
#> 95        NA    0.0                   16     MWI   interpolateBlues   0 1.0e-02
#> 96       100    0,0                   17     MWI   interpolateBlues   0 4.0e+04
#> 97       100    0,0                   17     MWI   interpolateBlues   0 4.0e+04
#> 98       100    0,0                   17     MWI   interpolateBlues   0 4.0e+04
#> 99        NA   0.0%                    9 default   interpolateBlues   0 1.0e+00
#> 100       NA   0.0%                    9 default   interpolateBlues   0 1.0e+00
#> 101       NA   0.0%                    9 default   interpolateBlues   0 1.0e+00
#> 102       10    0,0                   10 default   interpolateBlues   0 8.0e+04
#> 103       10      0                   10 default   interpolateBlues   0 8.0e+04
#> 104       10    0,0                   10 default   interpolateBlues   0 8.0e+04
#> 105       10      0                   11 default   interpolateBlues   0 8.0e+04
#> 106       10    0,0                   11 default   interpolateBlues   0 8.0e+04
#> 107       10    0,0                    8 default  interpolateGreens   0 8.0e+04
#> 108       10    0,0                    8 default  interpolateGreens   0 8.0e+04
#> 109      100    0,0                   12 default    interpolateReds   0 1.0e+07
#> 110      100    0,0                   12 default    interpolateReds   0 1.0e+07
#> 111       10    0,0                   13 default  interpolateGreens   0 8.0e+04
#> 112       10    0,0                   13 default  interpolateGreens   0 8.0e+04
#> 113       10    0,0                   14 default   interpolateBlues   0 8.0e+04
#> 114       10    0,0                   14 default   interpolateBlues   0 8.0e+04
#> 115       10    0,0                   15 default   interpolateBlues   0 8.0e+04
#> 116       10    0,0                   15 default   interpolateBlues   0 8.0e+04
#> 117       10    0,0                   21 default interpolateOranges   0 3.5e+05
#> 118       10    0,0                   21 default interpolateOranges   0 3.5e+05
#> 119       10    0,0                   23 default interpolateOranges   0 3.5e+05
#> 120       10    0,0                   23 default interpolateOranges   0 3.5e+05
#> 121       10    0,0                   27 default interpolateOranges   0 7.0e+05
#> 122       10    0,0                   27 default interpolateOranges   0 7.0e+05
#> 123       NA    0.0                   28 default   interpolateGreys   0 1.2e+00
#> 124       NA    0.0                   29 default    interpolateReds   0 1.2e+00
#> 125       NA    0.0                   30 default   interpolateBlues   0 1.2e+00
#> 126       NA    0.0                   31 default  interpolateGreens   0 1.2e+00
#> 127       NA    0.0                   32 default   interpolateBlues   0 1.2e+00
#> 128       NA    0.0                   33 default    interpolateReds   0 1.2e+00
#> 129       NA    0.0                   34 default  interpolateGreens   0 1.2e+00
#> 130       NA    0.0                   35 default   interpolateBlues   0 1.2e+00
#> 131       NA    0.0                   36 default   interpolateBlues   0 1.2e+00
#> 132       NA   0.0%                   37 default interpolateOranges   0 5.0e-01
#> 133       NA   0.0%                   38 default interpolateOranges   0 1.0e+00
#> 134       NA    0,0                    7 default interpolateOranges   0 1.0e+06
#> 135        1   0.0%                   39 default   interpolateGreys   0 1.2e+00
#>     invert_scale
#> 1          FALSE
#> 2          FALSE
#> 3          FALSE
#> 4          FALSE
#> 5          FALSE
#> 6          FALSE
#> 7          FALSE
#> 8          FALSE
#> 9          FALSE
#> 10         FALSE
#> 11         FALSE
#> 12         FALSE
#> 13         FALSE
#> 14         FALSE
#> 15         FALSE
#> 16         FALSE
#> 17         FALSE
#> 18         FALSE
#> 19         FALSE
#> 20         FALSE
#> 21         FALSE
#> 22         FALSE
#> 23         FALSE
#> 24         FALSE
#> 25         FALSE
#> 26         FALSE
#> 27         FALSE
#> 28         FALSE
#> 29         FALSE
#> 30         FALSE
#> 31         FALSE
#> 32         FALSE
#> 33         FALSE
#> 34         FALSE
#> 35         FALSE
#> 36         FALSE
#> 37         FALSE
#> 38         FALSE
#> 39         FALSE
#> 40         FALSE
#> 41         FALSE
#> 42         FALSE
#> 43         FALSE
#> 44         FALSE
#> 45         FALSE
#> 46         FALSE
#> 47         FALSE
#> 48         FALSE
#> 49         FALSE
#> 50         FALSE
#> 51         FALSE
#> 52         FALSE
#> 53         FALSE
#> 54         FALSE
#> 55         FALSE
#> 56         FALSE
#> 57         FALSE
#> 58         FALSE
#> 59         FALSE
#> 60         FALSE
#> 61         FALSE
#> 62         FALSE
#> 63         FALSE
#> 64         FALSE
#> 65         FALSE
#> 66         FALSE
#> 67         FALSE
#> 68         FALSE
#> 69         FALSE
#> 70         FALSE
#> 71         FALSE
#> 72         FALSE
#> 73         FALSE
#> 74         FALSE
#> 75         FALSE
#> 76         FALSE
#> 77         FALSE
#> 78         FALSE
#> 79         FALSE
#> 80         FALSE
#> 81         FALSE
#> 82         FALSE
#> 83         FALSE
#> 84         FALSE
#> 85         FALSE
#> 86         FALSE
#> 87         FALSE
#> 88         FALSE
#> 89         FALSE
#> 90         FALSE
#> 91         FALSE
#> 92         FALSE
#> 93         FALSE
#> 94         FALSE
#> 95         FALSE
#> 96         FALSE
#> 97         FALSE
#> 98         FALSE
#> 99         FALSE
#> 100        FALSE
#> 101        FALSE
#> 102        FALSE
#> 103        FALSE
#> 104        FALSE
#> 105        FALSE
#> 106        FALSE
#> 107        FALSE
#> 108        FALSE
#> 109        FALSE
#> 110        FALSE
#> 111        FALSE
#> 112        FALSE
#> 113        FALSE
#> 114        FALSE
#> 115        FALSE
#> 116        FALSE
#> 117        FALSE
#> 118        FALSE
#> 119        FALSE
#> 120        FALSE
#> 121        FALSE
#> 122        FALSE
#> 123        FALSE
#> 124        FALSE
#> 125        FALSE
#> 126        FALSE
#> 127        FALSE
#> 128        FALSE
#> 129        FALSE
#> 130        FALSE
#> 131        FALSE
#> 132        FALSE
#> 133        FALSE
#> 134        FALSE
#> 135        FALSE