Create output map plot

Create output map plot

map_outputs(
  naomi_geom,
  indicator,
  calendar_quarter,
  age,
  sex_disag,
  level = NULL,
  colour_palette = "red",
  fig_title = NULL,
  legend_title = NULL,
  legend_label = ggplot2::waiver(),
  breaks = 6
)

Arguments

naomi_geom

Labelled naomi output object with geometry

indicator

The indicator to generate plot for

calendar_quarter

Naomi projection quarter to filter supplied data to.

age

The age group to include in plots

sex_disag

Sex diaggregation to use in plot,

level

Integer, corresponding to desired area level to filter supplied data to. Default is lowest area level available in supplied data.

colour_palette

Names or hexcode value for colour scale to be applied to to lowest and highest value in the supplied data. Can be specified as a preset colour palette using "red", "green" and "blue" or manually as a named list: cols = c( start_colour = "red", end_colour = "blue").

fig_title

Title for the figure, blank by default

legend_title

Title for the legend, blank by default

legend_label

Legend label

breaks

Number of break points to create in scale

Value

Mapped outputs