Construct Model Frames and Adjacency Structures

Construct Model Frames and Adjacency Structures

naomi_model_frame(
  area_merged,
  population_agesex,
  spec,
  scope = areas$tree$area_id,
  level = max(areas$tree$Get("area_level")),
  calendar_quarter1,
  calendar_quarter2,
  calendar_quarter3,
  age_groups = get_five_year_age_groups(),
  sexes = c("male", "female"),
  omega = 0.7,
  rita_param = list(OmegaT0 = 130/365, sigma_OmegaT = ((142 - 118)/365)/(2 *
    qnorm(0.975)), betaT0 = 0, sigma_betaT = 1e-05, ritaT = 1),
  sigma_u_sd = 1,
  artattend = TRUE,
  artattend_t2 = FALSE,
  artattend_log_gamma_offset = -4,
  rho_paed_15to49f_ratio = FALSE,
  rho_paed_x_term = FALSE,
  logit_nu_mean = 2,
  logit_nu_sd = 0.3,
  spectrum_population_calibration = "national",
  output_aware_plhiv = TRUE
)

Arguments

area_merged
Merged version of area hierarchy

population_agesex
Population by age group and sex

spec
Spec

scope
The collection of area IDs to be modelled. Defaults to all area ids.

level
Admin level

calendar_quarter1
Calendar quarter at time 1 ("CYyyyyQq")

calendar_quarter2
Calendar quarter at time 2 ("CYyyyyQq")

calendar_quarter3
Calendar quarter at time 3 ("CYyyyyQq")

age_groups
Age groups to include in model frame

sexes
Sexes

omega
Omega

rita_param
rita_param

sigma_u_sd
sigma_u_sd

artattend
logical; whether to estimate neighboring district ART attendance

artattend_t2
logical; whether to allow time-varying neighboring district ART attendance

artattend_log_gamma_offset
logit offset for neigboring district ART attendance

rho_paed_15to49f_ratio
logical; to model paediatric prevalence as ratio of 15-49 female prevalence

rho_paed_x_term
logical; to include area interaction for paediatric prevalence

logit_nu_mean
mean of logit viral load suppression.

logit_nu_sd
standard deviation of logit viral load suppression.

spectrum_population_calibration
character string values "national", "subnational", "none"

Value

Naomi model frame

Details

Argument spectrum_population_calibration determines whether to calibrate population inputs to match Spectrum population by age and sex. If the Spectrum file is a single national Spectrum file, then options "national" and "subnational" return the same results.