Export proportion aware by five year age group from Shiny90

Export estimates for proportion aware of status from a Shiny90 output file to five-year age groups 15-19 to 50+.

extract_shiny90_age_sex(shiny90_path, pjnz_path = NULL, years = NULL)

Arguments

shiny90_path

file path to .shiny90 digest file.

pjnz_path

file path to PJNZ file

years

year(s) to generate estimates; an integer or a vector of integers. If NULL, all years available in estimates are returned (default).

Value

A data frame consisting of the number of PLHIV, aware of status and on ART by sex and five year age group 15-49 and age 50+ for specified years.

Details

The 'artnum' divided by 'plhiv' columns in the output will give a different ART coverage than Spectrum output for the same year because these outputs are based on the internal mid-year ART coverage in Spectrum, while Spectrum reports an end of year (Dec 31) ART coverage.

Examples

pjnz <- system.file("extdata/demo_mwi2019.PJNZ", package = "naomi")
shiny90dir <- tempfile()
utils::unzip(pjnz, "malawi.zip.shiny90", exdir = shiny90dir)
shiny90_path <- file.path(shiny90dir, "malawi.zip.shiny90")

extract_shiny90_age_sex(shiny90_path, pjnz, year = 2010:2019)
#>       area year    sex agegr hivstatus     plhiv      aware     artnum
#> 1   Malawi 2010 female 15-19  positive  27579.09  15253.635   4053.864
#> 2   Malawi 2011 female 15-19  positive  28094.73  16266.981   5766.776
#> 3   Malawi 2012 female 15-19  positive  28535.11  17260.553   8060.872
#> 4   Malawi 2013 female 15-19  positive  28835.70  17475.890  10218.728
#> 5   Malawi 2014 female 15-19  positive  29035.27  18055.328  12032.654
#> 6   Malawi 2015 female 15-19  positive  29170.52  18613.605  13400.573
#> 7   Malawi 2016 female 15-19  positive  29236.05  19421.099  15274.894
#> 8   Malawi 2017 female 15-19  positive  29049.73  21318.448  17245.598
#> 9   Malawi 2018 female 15-19  positive  28654.68  22396.728  18760.047
#> 10  Malawi 2019 female 15-19  positive  28070.21  22626.096  19376.252
#> 11  Malawi 2010   male 15-19  positive  11630.57   6143.187   2970.866
#> 12  Malawi 2011   male 15-19  positive  12656.24   7019.739   3893.862
#> 13  Malawi 2012   male 15-19  positive  13557.57   7794.975   4926.789
#> 14  Malawi 2013   male 15-19  positive  14327.54   8480.104   5951.973
#> 15  Malawi 2014   male 15-19  positive  14977.69   9277.140   6931.252
#> 16  Malawi 2015   male 15-19  positive  15545.79  10062.635   7867.091
#> 17  Malawi 2016   male 15-19  positive  16016.32  10835.561   8769.070
#> 18  Malawi 2017   male 15-19  positive  16332.45  11880.023   9818.548
#> 19  Malawi 2018   male 15-19  positive  16522.09  12805.906  10874.816
#> 20  Malawi 2019   male 15-19  positive  16609.83  13439.188  11304.906
#> 21  Malawi 2010 female 20-24  positive  55463.39  37377.056   6900.539
#> 22  Malawi 2011 female 20-24  positive  55985.84  39784.873  10326.225
#> 23  Malawi 2012 female 20-24  positive  56182.13  41475.685  15026.217
#> 24  Malawi 2013 female 20-24  positive  55950.38  41641.220  19614.124
#> 25  Malawi 2014 female 20-24  positive  55422.71  41929.802  23209.420
#> 26  Malawi 2015 female 20-24  positive  54796.30  42094.171  25917.423
#> 27  Malawi 2016 female 20-24  positive  54211.83  42662.849  29489.296
#> 28  Malawi 2017 female 20-24  positive  53551.88  44812.787  33436.383
#> 29  Malawi 2018 female 20-24  positive  52751.66  45920.643  36484.983
#> 30  Malawi 2019 female 20-24  positive  51740.54  45938.659  37964.230
#> 31  Malawi 2010   male 20-24  positive  19372.39   8875.105   1596.891
#> 32  Malawi 2011   male 20-24  positive  20393.48   9674.298   2283.138
#> 33  Malawi 2012   male 20-24  positive  21330.45  10360.152   3093.538
#> 34  Malawi 2013   male 20-24  positive  22062.95  10846.052   3928.129
#> 35  Malawi 2014   male 20-24  positive  22624.89  11513.197   5058.470
#> 36  Malawi 2015   male 20-24  positive  23081.34  12205.378   6075.080
#> 37  Malawi 2016   male 20-24  positive  23489.69  13079.759   7494.158
#> 38  Malawi 2017   male 20-24  positive  23781.15  14685.560   9214.891
#> 39  Malawi 2018   male 20-24  positive  23929.19  16021.815  10857.161
#> 40  Malawi 2019   male 20-24  positive  23923.85  16881.397  11404.376
#> 41  Malawi 2010 female 25-29  positive  72495.74  55503.310  14728.907
#> 42  Malawi 2011 female 25-29  positive  71934.59  57882.045  19039.944
#> 43  Malawi 2012 female 25-29  positive  72217.30  59989.606  25304.247
#> 44  Malawi 2013 female 25-29  positive  73177.97  61358.519  31638.408
#> 45  Malawi 2014 female 25-29  positive  74443.49  63168.116  37252.619
#> 46  Malawi 2015 female 25-29  positive  75548.33  64794.495  41977.835
#> 47  Malawi 2016 female 25-29  positive  76094.62  66313.541  47239.794
#> 48  Malawi 2017 female 25-29  positive  75783.24  68610.809  52573.370
#> 49  Malawi 2018 female 25-29  positive  74663.88  69267.693  56357.329
#> 50  Malawi 2019 female 25-29  positive  72985.16  68538.376  57949.173
#> 51  Malawi 2010   male 25-29  positive  37465.58  23574.973   4168.180
#> 52  Malawi 2011   male 25-29  positive  37104.15  23668.697   5136.606
#> 53  Malawi 2012   male 25-29  positive  37269.89  23700.814   6241.072
#> 54  Malawi 2013   male 25-29  positive  37840.98  23697.544   7307.139
#> 55  Malawi 2014   male 25-29  positive  38647.98  24321.370   8950.160
#> 56  Malawi 2015   male 25-29  positive  39460.89  25160.772  10480.476
#> 57  Malawi 2016   male 25-29  positive  40064.57  26332.554  12648.958
#> 58  Malawi 2017   male 25-29  positive  40226.24  28706.147  15408.740
#> 59  Malawi 2018   male 25-29  positive  39931.54  30359.626  17993.112
#> 60  Malawi 2019   male 25-29  positive  39294.59  31045.566  18614.587
#> 61  Malawi 2010 female 30-34  positive  85393.96  69332.900  22813.565
#> 62  Malawi 2011 female 30-34  positive  84443.26  72204.436  28238.383
#> 63  Malawi 2012 female 30-34  positive  83259.98  73536.531  35278.387
#> 64  Malawi 2013 female 30-34  positive  82071.11  73473.628  41557.817
#> 65  Malawi 2014 female 30-34  positive  81093.99  73425.087  46669.729
#> 66  Malawi 2015 female 30-34  positive  80560.41  73541.591  50836.715
#> 67  Malawi 2016 female 30-34  positive  80693.90  74398.300  55809.926
#> 68  Malawi 2017 female 30-34  positive  81460.99  76817.560  61598.820
#> 69  Malawi 2018 female 30-34  positive  82696.55  79140.071  66880.238
#> 70  Malawi 2019 female 30-34  positive  83985.19  80972.603  70717.683
#> 71  Malawi 2010   male 30-34  positive  56197.45  40671.815   9193.746
#> 72  Malawi 2011   male 30-34  positive  54778.06  40834.673  10975.563
#> 73  Malawi 2012   male 30-34  positive  53314.31  40136.587  12698.368
#> 74  Malawi 2013   male 30-34  positive  51923.10  38920.780  14025.011
#> 75  Malawi 2014   male 30-34  positive  50745.69  38108.645  15608.444
#> 76  Malawi 2015   male 30-34  positive  49945.28  37658.224  17013.997
#> 77  Malawi 2016   male 30-34  positive  49661.31  37956.502  19123.818
#> 78  Malawi 2017   male 30-34  positive  49849.62  39970.648  22298.007
#> 79  Malawi 2018   male 30-34  positive  50383.67  42078.981  25733.288
#> 80  Malawi 2019   male 30-34  positive  51009.49  43725.648  27168.447
#> 81  Malawi 2010 female 35-39  positive  78416.92  57970.406  27390.123
#> 82  Malawi 2011 female 35-39  positive  80985.86  65129.389  34568.236
#> 83  Malawi 2012 female 35-39  positive  83290.47  70845.975  43601.244
#> 84  Malawi 2013 female 35-39  positive  85195.24  74802.369  51674.331
#> 85  Malawi 2014 female 35-39  positive  86504.30  77680.230  58029.243
#> 86  Malawi 2015 female 35-39  positive  87113.38  79435.150  62949.635
#> 87  Malawi 2016 female 35-39  positive  87014.96  80384.572  67599.964
#> 88  Malawi 2017 female 35-39  positive  86325.27  81244.642  71755.298
#> 89  Malawi 2018 female 35-39  positive  85345.82  81482.416  74483.637
#> 90  Malawi 2019 female 35-39  positive  84396.97  81288.116  75674.323
#> 91  Malawi 2010   male 35-39  positive  61149.97  39903.583  15013.136
#> 92  Malawi 2011   male 35-39  positive  61364.79  42969.542  18228.471
#> 93  Malawi 2012   male 35-39  positive  61487.42  44940.143  21425.208
#> 94  Malawi 2013   male 35-39  positive  61459.43  46018.405  24046.947
#> 95  Malawi 2014   male 35-39  positive  61152.29  46636.396  26318.055
#> 96  Malawi 2015   male 35-39  positive  60503.34  46778.548  28183.484
#> 97  Malawi 2016   male 35-39  positive  59498.18  46690.159  30275.433
#> 98  Malawi 2017   male 35-39  positive  58212.91  47104.112  33081.290
#> 99  Malawi 2018   male 35-39  positive  56848.89  47462.908  35624.539
#> 100 Malawi 2019   male 35-39  positive  55620.28  47571.059  35807.994
#> 101 Malawi 2010 female 40-44  positive  55510.28  39666.727  21898.055
#> 102 Malawi 2011 female 40-44  positive  59570.68  46700.232  28283.669
#> 103 Malawi 2012 female 40-44  positive  63739.71  53324.343  36467.686
#> 104 Malawi 2013 female 40-44  positive  67950.22  59130.985  44466.104
#> 105 Malawi 2014 female 40-44  positive  72183.97  64684.062  51778.125
#> 106 Malawi 2015 female 40-44  positive  76724.03  70190.596  58917.452
#> 107 Malawi 2016 female 40-44  positive  80003.98  74436.479  65525.961
#> 108 Malawi 2017 female 40-44  positive  82746.73  78512.858  71791.016
#> 109 Malawi 2018 female 40-44  positive  84851.68  81655.711  76673.528
#> 110 Malawi 2019 female 40-44  positive  86214.41  83684.860  79637.919
#> 111 Malawi 2010   male 40-44  positive  50033.40  31878.106  14999.670
#> 112 Malawi 2011   male 40-44  positive  52120.28  35964.130  18817.041
#> 113 Malawi 2012   male 40-44  positive  54071.30  39409.555  22811.722
#> 114 Malawi 2013   male 40-44  positive  55858.31  42241.555  26463.131
#> 115 Malawi 2014   male 40-44  positive  57520.80  44844.029  29701.325
#> 116 Malawi 2015   male 40-44  positive  59462.18  47459.616  33018.203
#> 117 Malawi 2016   male 40-44  positive  60308.73  49183.830  36072.384
#> 118 Malawi 2017   male 40-44  positive  60840.11  51190.232  39802.780
#> 119 Malawi 2018   male 40-44  positive  61041.94  52825.382  43202.863
#> 120 Malawi 2019   male 40-44  positive  60879.46  53785.178  44104.990
#> 121 Malawi 2010 female 45-49  positive  34378.16  24369.410  15047.543
#> 122 Malawi 2011 female 45-49  positive  37889.81  29513.906  19585.940
#> 123 Malawi 2012 female 45-49  positive  41745.18  34782.239  25625.717
#> 124 Malawi 2013 female 45-49  positive  45807.17  39757.184  31840.553
#> 125 Malawi 2014 female 45-49  positive  49831.72  44628.004  37445.140
#> 126 Malawi 2015 female 45-49  positive  53337.78  48874.726  42594.634
#> 127 Malawi 2016 female 45-49  positive  57709.41  53890.042  48801.108
#> 128 Malawi 2017 female 45-49  positive  61990.11  59092.412  55139.247
#> 129 Malawi 2018 female 45-49  positive  66138.50  63959.770  60948.151
#> 130 Malawi 2019 female 45-49  positive  70206.87  68497.592  65899.155
#> 131 Malawi 2010   male 45-49  positive  33894.66  21680.343  11865.291
#> 132 Malawi 2011   male 45-49  positive  36299.89  25253.734  15104.103
#> 133 Malawi 2012   male 45-49  positive  38964.07  28756.359  18798.869
#> 134 Malawi 2013   male 45-49  positive  41738.11  32113.714  22542.850
#> 135 Malawi 2014   male 45-49  positive  44341.32  35352.014  25832.338
#> 136 Malawi 2015   male 45-49  positive  46231.01  37931.670  28806.680
#> 137 Malawi 2016   male 45-49  positive  48685.30  40978.048  32369.856
#> 138 Malawi 2017   male 45-49  positive  50834.52  44137.672  36513.892
#> 139 Malawi 2018   male 45-49  positive  52683.95  46975.387  40435.383
#> 140 Malawi 2019   male 45-49  positive  54372.78  49408.897  42532.876
#> 141 Malawi 2010 female 50-99  positive  46218.22  30026.528  21504.405
#> 142 Malawi 2011 female 50-99  positive  50426.22  36665.273  27452.327
#> 143 Malawi 2012 female 50-99  positive  55300.63  43627.777  35278.131
#> 144 Malawi 2013 female 50-99  positive  60896.41  50562.895  43557.370
#> 145 Malawi 2014 female 50-99  positive  67211.39  58049.392  51672.920
#> 146 Malawi 2015 female 50-99  positive  74243.41  66082.528  60379.820
#> 147 Malawi 2016 female 50-99  positive  81987.38  74841.171  70418.377
#> 148 Malawi 2017 female 50-99  positive  90368.15  84681.225  81368.483
#> 149 Malawi 2018 female 50-99  positive  99294.97  94857.803  92390.749
#> 150 Malawi 2019 female 50-99  positive 108597.84 105035.750 102777.636
#> 151 Malawi 2010   male 50-99  positive  43231.82  25618.095  16824.221
#> 152 Malawi 2011   male 50-99  positive  46772.88  30682.782  21549.715
#> 153 Malawi 2012   male 50-99  positive  50827.58  35909.359  27013.934
#> 154 Malawi 2013   male 50-99  positive  55392.50  41330.955  32869.319
#> 155 Malawi 2014   male 50-99  positive  60410.01  47203.143  38558.457
#> 156 Malawi 2015   male 50-99  positive  65869.35  53436.650  44812.989
#> 157 Malawi 2016   male 50-99  positive  71758.62  60181.738  51932.823
#> 158 Malawi 2017   male 50-99  positive  78008.16  67795.087  60483.853
#> 159 Malawi 2018   male 50-99  positive  84509.40  75664.915  69353.338
#> 160 Malawi 2019   male 50-99  positive  91127.06  83356.992  76074.825