Get indicator metadata

Get indicator metadata

get_meta_indicator()

Value

data.frame of indicator ids, labels, descriptions, and parameter mapping.

Examples

get_meta_indicator()
#>                       indicator                      indicator_label
#> 1                    population                           Population
#> 2                    prevalence                       HIV prevalence
#> 3                         plhiv                                PLHIV
#> 4                  art_coverage                         ART coverage
#> 5         art_current_residents               ART number (residents)
#> 6                   art_current               ART number (attending)
#> 7                 number_on_art                        Number on ART
#> 8           untreated_plhiv_num                     PLHIV not on ART
#> 9              aware_plhiv_prop               Proportion PLHIV aware
#> 10            unaware_plhiv_num                 Number PLHIV unaware
#> 11              aware_plhiv_num                   Number PLHIV aware
#> 12                 plhiv_attend                PLHIV (ART catchment)
#> 13       untreated_plhiv_attend      Untreated PLHIV (ART catchment)
#> 14           aware_plhiv_attend   Number Aware PLHIV (ART catchment)
#> 15         unaware_plhiv_attend Number unaware PLHIV (ART catchment)
#> 16                    incidence                        HIV incidence
#> 17                   infections                       New infections
#> 18               anc_prevalence                   ANC HIV prevalence
#> 19             anc_art_coverage               ANC prior ART coverage
#> 20                  anc_clients                          ANC clients
#> 21                    anc_plhiv           HIV positive ANC attendees
#> 22              anc_already_art         ANC attendees already on ART
#> 23                  anc_art_new               ART initiations at ANC
#> 24                anc_known_pos                   ANC known positive
#> 25                   anc_tested                           ANC tested
#> 26               anc_tested_pos                  ANC tested positive
#> 27               anc_tested_neg                  ANC tested negative
#> 28             population_ratio                     Population ratio
#> 29                  plhiv_ratio                          PLHIV ratio
#> 30             infections_ratio                 New infections ratio
#> 31            art_current_ratio         ART number (attending) ratio
#> 32      unaware_plhiv_num_ratio           Number PLHIV unaware ratio
#> 33             prevalence_ratio                 HIV prevalence ratio
#> 34           art_coverage_ratio                   ART coverage ratio
#> 35       aware_plhiv_prop_ratio         Proportion PLHIV aware ratio
#> 36              incidence_ratio                      Incidence ratio
#> 37   anc_prevalence_age_matched           ANC prevalence age matched
#> 38 anc_art_coverage_age_matched         ANC ART coverage age matched
#> 39        population_proportion                Population proportion
#>                                                                      description
#> 1                                                                Population size
#> 2                                    Proportion of total population HIV positive
#> 3                                               Number of people living with HIV
#> 4                                         Proportion of PLHIV on ART (residents)
#> 5                                                      Number on ART (residents)
#> 6                                               Number receiving ART (attending)
#> 7                                                      Number on ART description
#> 8                                                     Number of PLHIV not on ART
#> 9                               Proportion of PLHIV aware of HIV positive status
#> 10                                   Number PLHIV unaware of HIV positive status
#> 11                                               INDICATOR_LABEL_AWARE_PLHIV_NUM
#> 12                                         Number of PLHIV in area ART catchment
#> 13                               Number of untreated PLHIV in area ART catchment
#> 14                     Number of PLHIV aware of HIV status in area ART catchment
#> 15                   Number of PLHIV unaware of HIV status in area ART catchment
#> 16                                                   HIV incidence rate per year
#> 17                                             Number of new infections per year
#> 18                                            HIV prevalence among ANC attendees
#> 19                           ART coverage among ANC attendees prior to first ANC
#> 20                                              Total number attending first ANC
#> 21                                    Number of HIV positive women attending ANC
#> 22                          Number already on ART prior prior to first ANC visit
#> 23                              Number of pregnant women initiated on ART at ANC
#> 24 Number of HIV positive ANC attendees aware of status prior to first ANC visit
#> 25                                          Total number of ANC attendees tested
#> 26                            Number of HIV positive ANC attendees tested at ANC
#> 27                            Number of HIV negative ANC attendees tested at ANC
#> 28                                                                            NA
#> 29                                                                            NA
#> 30                                                                            NA
#> 31                                                                            NA
#> 32                                                                            NA
#> 33                                                                            NA
#> 34                                                                            NA
#> 35                                                                            NA
#> 36                                                                            NA
#> 37                                                                            NA
#> 38                                                                            NA
#> 39                                                                            NA
#>                     parameter indicator_sort_order anc_indicator format scale
#> 1              population_out                    1         FALSE     NA    NA
#> 2                     rho_out                    2         FALSE     NA    NA
#> 3                   plhiv_out                    3         FALSE     NA    NA
#> 4                   alpha_out                    4         FALSE     NA    NA
#> 5                  artnum_out                    5         FALSE     NA    NA
#> 6               artattend_out                    6         FALSE     NA    NA
#> 7                        <NA>                    7          TRUE     NA    NA
#> 8     untreated_plhiv_num_out                    8         FALSE     NA    NA
#> 9        aware_plhiv_prop_out                    9         FALSE     NA    NA
#> 10      unaware_plhiv_num_out                   10         FALSE     NA    NA
#> 11        aware_plhiv_num_out                   11         FALSE     NA    NA
#> 12           plhiv_attend_out                   12         FALSE     NA    NA
#> 13 untreated_plhiv_attend_out                   13         FALSE     NA    NA
#> 14     aware_plhiv_attend_out                   14         FALSE     NA    NA
#> 15   unaware_plhiv_attend_out                   15         FALSE     NA    NA
#> 16                 lambda_out                   16         FALSE     NA    NA
#> 17             infections_out                   17         FALSE     NA    NA
#> 18                anc_rho_out                   18          TRUE     NA    NA
#> 19              anc_alpha_out                   19          TRUE     NA    NA
#> 20            anc_clients_out                   20          TRUE     NA    NA
#> 21              anc_plhiv_out                   21          TRUE     NA    NA
#> 22        anc_already_art_out                   22          TRUE     NA    NA
#> 23            anc_art_new_out                   23          TRUE     NA    NA
#> 24          anc_known_pos_out                   24          TRUE     NA    NA
#> 25                       <NA>                   25          TRUE     NA    NA
#> 26         anc_tested_pos_out                   26          TRUE     NA    NA
#> 27         anc_tested_neg_out                   27          TRUE     NA    NA
#> 28                       <NA>                   28         FALSE     NA    NA
#> 29                       <NA>                   29         FALSE     NA    NA
#> 30                       <NA>                   30         FALSE     NA    NA
#> 31                       <NA>                   31         FALSE     NA    NA
#> 32                       <NA>                   32         FALSE     NA    NA
#> 33                       <NA>                   33         FALSE     NA    NA
#> 34                       <NA>                   34         FALSE     NA    NA
#> 35                       <NA>                   35         FALSE     NA    NA
#> 36                       <NA>                   36         FALSE     NA    NA
#> 37                       <NA>                   37          TRUE     NA    NA
#> 38                       <NA>                   38          TRUE     NA    NA
#> 39                       <NA>                   39         FALSE     NA    NA