Get plotting metadata for a particular country

Get plotting metadata for a particular country

get_plotting_metadata(iso3)

Arguments

iso3
iso3 code of country to get metadata for or blank for default configuration.

Value

List of plotting metadata about how to locate data for a specific indicator, data type and plot type. Also returns metadata about colour scheme to use for that country and indicator.

Examples

get_plotting_metadata("MWI")

#>                  indicator data_type  plot_type     value_column
#> 1               prevalence    survey choropleth         estimate
#> 2               prevalence    output   barchart             mean
#> 3               prevalence    output choropleth             mean
#> 4             art_coverage    survey choropleth         estimate
#> 5             art_coverage    output choropleth             mean
#> 6             art_coverage    output   barchart             mean
#> 7              art_current    output   barchart             mean
#> 8              art_current    output choropleth             mean
#> 9              art_current programme choropleth      art_current
#> 10         recent_infected    survey choropleth         estimate
#> 11 viral_suppression_plhiv    survey choropleth         estimate
#> 12          anc_prevalence    output choropleth             mean
#> 13          anc_prevalence    output   barchart             mean
#> 14          anc_prevalence       anc choropleth   anc_prevalence
#> 15        anc_art_coverage       anc choropleth anc_art_coverage
#> 16        anc_art_coverage    output choropleth             mean
#> 17        anc_art_coverage    output   barchart             mean
#> 18   art_current_residents    output choropleth             mean
#> 19   art_current_residents    output   barchart             mean
#> 20              population    output choropleth             mean
#> 21              population    output   barchart             mean
#> 22                   plhiv    output   barchart             mean
#> 23                   plhiv    output choropleth             mean
#> 24               incidence    output choropleth             mean
#> 25               incidence    output   barchart             mean
#> 26              infections    output   barchart             mean
#> 27              infections    output choropleth             mean
#> 28        aware_plhiv_prop    output   barchart             mean
#> 29        aware_plhiv_prop    output choropleth             mean
#> 30       unaware_plhiv_num    output   barchart             mean
#> 31       unaware_plhiv_num    output choropleth             mean
#> 32         aware_plhiv_num    output choropleth             mean
#> 33         aware_plhiv_num    output   barchart             mean
#> 34     untreated_plhiv_num    output   barchart             mean
#> 35     untreated_plhiv_num    output choropleth             mean
#> 36             anc_clients    output choropleth             mean
#> 37             anc_clients    output   barchart             mean
#> 38               anc_plhiv    output choropleth             mean
#> 39               anc_plhiv    output   barchart             mean
#> 40         anc_already_art    output choropleth             mean
#> 41         anc_already_art    output   barchart             mean
#> 42             anc_art_new    output choropleth             mean
#> 43             anc_art_new    output   barchart             mean
#> 44           anc_known_pos    output choropleth             mean
#> 45           anc_known_pos    output   barchart             mean
#> 46          anc_tested_pos    output choropleth             mean
#> 47          anc_tested_pos    output   barchart             mean
#> 48          anc_tested_neg    output choropleth             mean
#> 49          anc_tested_neg    output   barchart             mean
#>    error_low_column error_high_column indicator_column         indicator_value
#> 1                                            indicator              prevalence
#> 2             lower             upper        indicator              prevalence
#> 3             lower             upper        indicator              prevalence
#> 4                                            indicator            art_coverage
#> 5             lower             upper        indicator            art_coverage
#> 6             lower             upper        indicator            art_coverage
#> 7             lower             upper        indicator             art_current
#> 8             lower             upper        indicator             art_current
#> 9                                                                             
#> 10                                           indicator         recent_infected
#> 11                                           indicator viral_suppression_plhiv
#> 12            lower             upper        indicator          anc_prevalence
#> 13            lower             upper        indicator          anc_prevalence
#> 14                                                                            
#> 15                                                                            
#> 16            lower             upper        indicator        anc_art_coverage
#> 17            lower             upper        indicator        anc_art_coverage
#> 18            lower             upper        indicator   art_current_residents
#> 19            lower             upper        indicator   art_current_residents
#> 20            lower             upper        indicator              population
#> 21            lower             upper        indicator              population
#> 22            lower             upper        indicator                   plhiv
#> 23            lower             upper        indicator                   plhiv
#> 24            lower             upper        indicator               incidence
#> 25            lower             upper        indicator               incidence
#> 26            lower             upper        indicator              infections
#> 27            lower             upper        indicator              infections
#> 28            lower             upper        indicator        aware_plhiv_prop
#> 29            lower             upper        indicator        aware_plhiv_prop
#> 30            lower             upper        indicator       unaware_plhiv_num
#> 31            lower             upper        indicator       unaware_plhiv_num
#> 32            lower             upper        indicator         aware_plhiv_num
#> 33            lower             upper        indicator         aware_plhiv_num
#> 34            lower             upper        indicator     untreated_plhiv_num
#> 35            lower             upper        indicator     untreated_plhiv_num
#> 36            lower             upper        indicator             anc_clients
#> 37            lower             upper        indicator             anc_clients
#> 38            lower             upper        indicator               anc_plhiv
#> 39            lower             upper        indicator               anc_plhiv
#> 40            lower             upper        indicator         anc_already_art
#> 41            lower             upper        indicator         anc_already_art
#> 42            lower             upper        indicator             anc_art_new
#> 43            lower             upper        indicator             anc_art_new
#> 44            lower             upper        indicator           anc_known_pos
#> 45            lower             upper        indicator           anc_known_pos
#> 46            lower             upper        indicator          anc_tested_pos
#> 47            lower             upper        indicator          anc_tested_pos
#> 48            lower             upper        indicator          anc_tested_neg
#> 49            lower             upper        indicator          anc_tested_neg
#>                            name scale accuracy format indicator_sort_order
#> 1                HIV prevalence     1       NA   0.0%                    2
#> 2                HIV prevalence     1       NA   0.0%                    2
#> 3                HIV prevalence     1       NA   0.0%                    2
#> 4                  ART coverage     1       NA   0.0%                    4
#> 5                  ART coverage     1       NA   0.0%                    4
#> 6                  ART coverage     1       NA   0.0%                    4
#> 7        ART number (attending)     1      100      0                    6
#> 8        ART number (attending)     1      100      0                    6
#> 9        ART number (attending)     1      100      0                   NA
#> 10 Proportion recently infected     1        1  0.00%                   NA
#> 11       Viral load suppression     1       NA  0.00%                   NA
#> 12           ANC HIV prevalence     1       NA   0.0%                   13
#> 13           ANC HIV prevalence     1       NA   0.0%                   13
#> 14           ANC HIV prevalence     1       NA   0.0%                   NA
#> 15       ANC prior ART coverage     1       NA   0.0%                   NA
#> 16       ANC prior ART coverage     1       NA   0.0%                   14
#> 17       ANC prior ART coverage     1       NA   0.0%                   14
#> 18       ART number (residents)     1      100      0                    5
#> 19       ART number (residents)     1      100      0                    5
#> 20                   Population     1     1000      0                    1
#> 21                   Population     1     1000      0                    1
#> 22                        PLHIV     1      100      0                    3
#> 23                        PLHIV     1      100      0                    3
#> 24                HIV incidence  1000       NA      0                   11
#> 25                HIV incidence  1000       NA      0                   11
#> 26               New infections     1      100      0                   12
#> 27               New infections     1      100      0                   12
#> 28       Proportion PLHIV aware     1       NA   0.0%                    8
#> 29       Proportion PLHIV aware     1       NA   0.0%                    8
#> 30         Number PLHIV unaware     1       10      0                    9
#> 31         Number PLHIV unaware     1       10      0                    9
#> 32           Number PLHIV aware     1       10      0                   10
#> 33           Number PLHIV aware     1       10      0                   10
#> 34             PLHIV not on ART     1       10      0                    7
#> 35             PLHIV not on ART     1       10      0                    7
#> 36                  ANC clients     1      100      0                   15
#> 37                  ANC clients     1      100      0                   15
#> 38   HIV positive ANC attendees     1       10      0                   16
#> 39   HIV positive ANC attendees     1       10      0                   16
#> 40 ANC attendees already on ART     1       10      0                   17
#> 41 ANC attendees already on ART     1       10      0                   17
#> 42       ART initiations at ANC     1       10      0                   18
#> 43       ART initiations at ANC     1       10      0                   18
#> 44           ANC known positive     1       10      0                   19
#> 45           ANC known positive     1       10      0                   19
#> 46          ANC tested positive     1       10      0                   20
#> 47          ANC tested positive     1       10      0                   20
#> 48          ANC tested negative     1       10      0                   21
#> 49          ANC tested negative     1       10      0                   21
#>       iso3             colour min     max invert_scale
#> 1      MWI   interpolateMagma   0 5.0e-01         TRUE
#> 2      MWI   interpolateMagma   0 5.0e-01         TRUE
#> 3      MWI   interpolateMagma   0 5.0e-01         TRUE
#> 4      MWI interpolateViridis   0 1.0e+00        FALSE
#> 5      MWI interpolateViridis   0 1.0e+00        FALSE
#> 6      MWI interpolateViridis   0 1.0e+00        FALSE
#> 7      MWI interpolateViridis   0 8.0e+04        FALSE
#> 8      MWI interpolateViridis   0 8.0e+04        FALSE
#> 9      MWI interpolateViridis   0 8.0e+04        FALSE
#> 10     MWI  interpolatePlasma   0 6.0e-01         TRUE
#> 11     MWI interpolateViridis   0 1.0e+00        FALSE
#> 12     MWI   interpolateMagma   0 5.0e-01         TRUE
#> 13     MWI   interpolateMagma   0 5.0e-01         TRUE
#> 14     MWI   interpolateMagma   0 5.0e-01         TRUE
#> 15     MWI interpolateViridis   0 1.0e+00        FALSE
#> 16     MWI interpolateViridis   0 1.0e+00        FALSE
#> 17     MWI interpolateViridis   0 1.0e+00        FALSE
#> 18     MWI interpolateViridis   0 7.0e+04        FALSE
#> 19     MWI interpolateViridis   0 7.0e+04        FALSE
#> 20     MWI interpolateViridis   0 2.0e+07        FALSE
#> 21     MWI interpolateViridis   0 2.0e+07        FALSE
#> 22     MWI   interpolateMagma   0 1.0e+07         TRUE
#> 23     MWI   interpolateMagma   0 1.0e+07         TRUE
#> 24     MWI   interpolateMagma   0 1.0e-02         TRUE
#> 25     MWI   interpolateMagma   0 1.0e-02         TRUE
#> 26     MWI   interpolateMagma   0 4.0e+04         TRUE
#> 27     MWI   interpolateMagma   0 4.0e+04         TRUE
#> 28 default interpolateViridis   0 1.0e+00        FALSE
#> 29 default interpolateViridis   0 1.0e+00        FALSE
#> 30 default interpolateViridis   0 8.0e+04         TRUE
#> 31 default interpolateViridis   0 8.0e+04         TRUE
#> 32 default interpolateViridis   0 8.0e+04         TRUE
#> 33 default interpolateViridis   0 8.0e+04         TRUE
#> 34 default interpolateViridis   0 8.0e+04         TRUE
#> 35 default interpolateViridis   0 8.0e+04         TRUE
#> 36 default interpolateViridis   0 7.0e+05        FALSE
#> 37 default interpolateViridis   0 7.0e+05        FALSE
#> 38 default   interpolateMagma   0 3.5e+05         TRUE
#> 39 default   interpolateMagma   0 3.5e+05         TRUE
#> 40 default interpolateViridis   0 3.5e+05        FALSE
#> 41 default interpolateViridis   0 3.5e+05        FALSE
#> 42 default interpolateViridis   0 3.5e+05        FALSE
#> 43 default interpolateViridis   0 3.5e+05        FALSE
#> 44 default   interpolateMagma   0 3.5e+05         TRUE
#> 45 default   interpolateMagma   0 3.5e+05         TRUE
#> 46 default   interpolateMagma   0 3.5e+05         TRUE
#> 47 default   interpolateMagma   0 3.5e+05         TRUE
#> 48 default interpolateViridis   0 7.0e+05        FALSE
#> 49 default interpolateViridis   0 7.0e+05        FALSE