Get plotting metadata for a particular country

Get plotting metadata for a particular country

get_plotting_metadata(iso3)

Arguments

iso3

iso3 code of country to get metadata for or blank for default configuration.

Value

List of plotting metadata about how to locate data for a specific indicator, data type and plot type. Also returns metadata about colour scheme to use for that country and indicator.

Examples

get_plotting_metadata("MWI")
#>                       indicator  data_type  plot_type        value_column
#> 1                    prevalence     survey choropleth            estimate
#> 2                    prevalence     output   barchart                mean
#> 3                    prevalence     output choropleth                mean
#> 4                    prevalence  calibrate   barchart                mean
#> 5                    prevalence comparison   barchart                mean
#> 6                  art_coverage     output choropleth                mean
#> 7                  art_coverage     output   barchart                mean
#> 8                  art_coverage     survey choropleth            estimate
#> 9                  art_coverage  calibrate   barchart                mean
#> 10                 art_coverage comparison   barchart                mean
#> 11                  art_current     output   barchart                mean
#> 12                  art_current  calibrate   barchart                mean
#> 13                  art_current  programme choropleth         art_current
#> 14                  art_current comparison   barchart                mean
#> 15                  art_current     output choropleth                mean
#> 16                      art_new  programme choropleth             art_new
#> 17              vl_tested_12mos  programme choropleth     vl_tested_12mos
#> 18          vl_suppressed_12mos  programme choropleth vl_suppressed_12mos
#> 19              recent_infected     survey choropleth            estimate
#> 20      viral_suppression_plhiv     survey choropleth            estimate
#> 21               anc_prevalence        anc choropleth      anc_prevalence
#> 22               anc_prevalence     output   barchart                mean
#> 23               anc_prevalence     output choropleth                mean
#> 24             anc_art_coverage     output   barchart                mean
#> 25             anc_art_coverage        anc choropleth    anc_art_coverage
#> 26             anc_art_coverage     output choropleth                mean
#> 27        art_current_residents     output   barchart                mean
#> 28        art_current_residents     output choropleth                mean
#> 29                   population  calibrate   barchart                mean
#> 30                   population     output   barchart                mean
#> 31                   population     output choropleth                mean
#> 32                        plhiv     output choropleth                mean
#> 33                        plhiv  calibrate   barchart                mean
#> 34                        plhiv     output   barchart                mean
#> 35                    incidence     output choropleth                mean
#> 36                    incidence  calibrate   barchart                mean
#> 37                    incidence     output   barchart                mean
#> 38                   infections  calibrate   barchart                mean
#> 39                   infections     output   barchart                mean
#> 40                   infections     output choropleth                mean
#> 41             aware_plhiv_prop     output choropleth                mean
#> 42             aware_plhiv_prop     output   barchart                mean
#> 43             aware_plhiv_prop  calibrate   barchart                mean
#> 44            unaware_plhiv_num     output choropleth                mean
#> 45            unaware_plhiv_num     output   barchart                mean
#> 46            unaware_plhiv_num  calibrate   barchart                mean
#> 47              aware_plhiv_num     output choropleth                mean
#> 48              aware_plhiv_num     output   barchart                mean
#> 49          untreated_plhiv_num     output   barchart                mean
#> 50          untreated_plhiv_num     output choropleth                mean
#> 51                 plhiv_attend     output choropleth                mean
#> 52                 plhiv_attend     output   barchart                mean
#> 53       untreated_plhiv_attend     output   barchart                mean
#> 54       untreated_plhiv_attend     output choropleth                mean
#> 55           aware_plhiv_attend     output   barchart                mean
#> 56           aware_plhiv_attend     output choropleth                mean
#> 57         unaware_plhiv_attend     output choropleth                mean
#> 58         unaware_plhiv_attend     output   barchart                mean
#> 59                  anc_clients     output choropleth                mean
#> 60                  anc_clients     output   barchart                mean
#> 61                    anc_plhiv     output   barchart                mean
#> 62                    anc_plhiv     output choropleth                mean
#> 63              anc_already_art     output   barchart                mean
#> 64              anc_already_art     output choropleth                mean
#> 65                  anc_art_new     output choropleth                mean
#> 66                  anc_art_new     output   barchart                mean
#> 67                anc_known_pos     output choropleth                mean
#> 68                anc_known_pos     output   barchart                mean
#> 69               anc_tested_pos     output choropleth                mean
#> 70               anc_tested_pos     output   barchart                mean
#> 71               anc_tested_neg     output   barchart                mean
#> 72               anc_tested_neg     output choropleth                mean
#> 73             population_ratio  calibrate   barchart                mean
#> 74                  plhiv_ratio  calibrate   barchart                mean
#> 75             infections_ratio  calibrate   barchart                mean
#> 76            art_current_ratio  calibrate   barchart                mean
#> 77      unaware_plhiv_num_ratio  calibrate   barchart                mean
#> 78             prevalence_ratio  calibrate   barchart                mean
#> 79           art_coverage_ratio  calibrate   barchart                mean
#> 80       aware_plhiv_prop_ratio  calibrate   barchart                mean
#> 81              incidence_ratio  calibrate   barchart                mean
#> 82   anc_prevalence_age_matched comparison   barchart                mean
#> 83 anc_art_coverage_age_matched comparison   barchart                mean
#>    error_low_column error_high_column indicator_column
#> 1                                            indicator
#> 2             lower             upper        indicator
#> 3             lower             upper        indicator
#> 4             lower             upper        indicator
#> 5             lower             upper        indicator
#> 6             lower             upper        indicator
#> 7             lower             upper        indicator
#> 8                                            indicator
#> 9             lower             upper        indicator
#> 10            lower             upper        indicator
#> 11            lower             upper        indicator
#> 12            lower             upper        indicator
#> 13                                                    
#> 14            lower             upper        indicator
#> 15            lower             upper        indicator
#> 16                                                    
#> 17                                                    
#> 18                                                    
#> 19                                           indicator
#> 20                                           indicator
#> 21                                                    
#> 22            lower             upper        indicator
#> 23            lower             upper        indicator
#> 24            lower             upper        indicator
#> 25                                                    
#> 26            lower             upper        indicator
#> 27            lower             upper        indicator
#> 28            lower             upper        indicator
#> 29            lower             upper        indicator
#> 30            lower             upper        indicator
#> 31            lower             upper        indicator
#> 32            lower             upper        indicator
#> 33            lower             upper        indicator
#> 34            lower             upper        indicator
#> 35            lower             upper        indicator
#> 36            lower             upper        indicator
#> 37            lower             upper        indicator
#> 38            lower             upper        indicator
#> 39            lower             upper        indicator
#> 40            lower             upper        indicator
#> 41            lower             upper        indicator
#> 42            lower             upper        indicator
#> 43            lower             upper        indicator
#> 44            lower             upper        indicator
#> 45            lower             upper        indicator
#> 46            lower             upper        indicator
#> 47            lower             upper        indicator
#> 48            lower             upper        indicator
#> 49            lower             upper        indicator
#> 50            lower             upper        indicator
#> 51            lower             upper        indicator
#> 52            lower             upper        indicator
#> 53            lower             upper        indicator
#> 54            lower             upper        indicator
#> 55            lower             upper        indicator
#> 56            lower             upper        indicator
#> 57            lower             upper        indicator
#> 58            lower             upper        indicator
#> 59            lower             upper        indicator
#> 60            lower             upper        indicator
#> 61            lower             upper        indicator
#> 62            lower             upper        indicator
#> 63            lower             upper        indicator
#> 64            lower             upper        indicator
#> 65            lower             upper        indicator
#> 66            lower             upper        indicator
#> 67            lower             upper        indicator
#> 68            lower             upper        indicator
#> 69            lower             upper        indicator
#> 70            lower             upper        indicator
#> 71            lower             upper        indicator
#> 72            lower             upper        indicator
#> 73            lower             upper        indicator
#> 74            lower             upper        indicator
#> 75            lower             upper        indicator
#> 76            lower             upper        indicator
#> 77            lower             upper        indicator
#> 78            lower             upper        indicator
#> 79            lower             upper        indicator
#> 80            lower             upper        indicator
#> 81            lower             upper        indicator
#> 82            lower             upper        indicator
#> 83            lower             upper        indicator
#>                 indicator_value                                 name scale
#> 1                    prevalence                       HIV prevalence     1
#> 2                    prevalence                       HIV prevalence     1
#> 3                    prevalence                       HIV prevalence     1
#> 4                    prevalence                       HIV prevalence     1
#> 5                    prevalence                       HIV prevalence     1
#> 6                  art_coverage                         ART coverage     1
#> 7                  art_coverage                         ART coverage     1
#> 8                  art_coverage                         ART coverage     1
#> 9                  art_coverage                         ART coverage     1
#> 10                 art_coverage                         ART coverage     1
#> 11                  art_current               ART number (attending)     1
#> 12                  art_current               ART number (attending)     1
#> 13                                            ART number (attending)     1
#> 14                  art_current               ART number (attending)     1
#> 15                  art_current               ART number (attending)     1
#> 16                                                           ART new     1
#> 17                                                         VL tested     1
#> 18                                               VL tests suppressed     1
#> 19              recent_infected         Proportion recently infected     1
#> 20      viral_suppression_plhiv               Viral load suppression     1
#> 21                                                ANC HIV prevalence     1
#> 22               anc_prevalence                   ANC HIV prevalence     1
#> 23               anc_prevalence                   ANC HIV prevalence     1
#> 24             anc_art_coverage               ANC prior ART coverage     1
#> 25                                            ANC prior ART coverage     1
#> 26             anc_art_coverage               ANC prior ART coverage     1
#> 27        art_current_residents               ART number (residents)     1
#> 28        art_current_residents               ART number (residents)     1
#> 29                   population                           Population     1
#> 30                   population                           Population     1
#> 31                   population                           Population     1
#> 32                        plhiv                                PLHIV     1
#> 33                        plhiv                                PLHIV     1
#> 34                        plhiv                                PLHIV     1
#> 35                    incidence                        HIV incidence  1000
#> 36                    incidence                        HIV incidence  1000
#> 37                    incidence                        HIV incidence  1000
#> 38                   infections                       New infections     1
#> 39                   infections                       New infections     1
#> 40                   infections                       New infections     1
#> 41             aware_plhiv_prop               Proportion PLHIV aware     1
#> 42             aware_plhiv_prop               Proportion PLHIV aware     1
#> 43             aware_plhiv_prop               Proportion PLHIV aware     1
#> 44            unaware_plhiv_num                 Number PLHIV unaware     1
#> 45            unaware_plhiv_num                 Number PLHIV unaware     1
#> 46            unaware_plhiv_num                 Number PLHIV unaware     1
#> 47              aware_plhiv_num                   Number PLHIV aware     1
#> 48              aware_plhiv_num                   Number PLHIV aware     1
#> 49          untreated_plhiv_num                     PLHIV not on ART     1
#> 50          untreated_plhiv_num                     PLHIV not on ART     1
#> 51                 plhiv_attend                PLHIV (ART catchment)     1
#> 52                 plhiv_attend                PLHIV (ART catchment)     1
#> 53       untreated_plhiv_attend      Untreated PLHIV (ART catchment)     1
#> 54       untreated_plhiv_attend      Untreated PLHIV (ART catchment)     1
#> 55           aware_plhiv_attend   Number Aware PLHIV (ART catchment)     1
#> 56           aware_plhiv_attend   Number Aware PLHIV (ART catchment)     1
#> 57         unaware_plhiv_attend Number unaware PLHIV (ART catchment)     1
#> 58         unaware_plhiv_attend Number unaware PLHIV (ART catchment)     1
#> 59                  anc_clients                          ANC clients     1
#> 60                  anc_clients                          ANC clients     1
#> 61                    anc_plhiv           HIV positive ANC attendees     1
#> 62                    anc_plhiv           HIV positive ANC attendees     1
#> 63              anc_already_art         ANC attendees already on ART     1
#> 64              anc_already_art         ANC attendees already on ART     1
#> 65                  anc_art_new               ART initiations at ANC     1
#> 66                  anc_art_new               ART initiations at ANC     1
#> 67                anc_known_pos                   ANC known positive     1
#> 68                anc_known_pos                   ANC known positive     1
#> 69               anc_tested_pos                  ANC tested positive     1
#> 70               anc_tested_pos                  ANC tested positive     1
#> 71               anc_tested_neg                  ANC tested negative     1
#> 72               anc_tested_neg                  ANC tested negative     1
#> 73             population_ratio                     Population ratio     1
#> 74                  plhiv_ratio                          PLHIV ratio     1
#> 75             infections_ratio                 New infections ratio     1
#> 76            art_current_ratio         ART number (attending) ratio     1
#> 77      unaware_plhiv_num_ratio           Number PLHIV unaware ratio     1
#> 78             prevalence_ratio                 HIV prevalence ratio     1
#> 79           art_coverage_ratio                   ART coverage ratio     1
#> 80       aware_plhiv_prop_ratio         Proportion PLHIV aware ratio     1
#> 81              incidence_ratio                      Incidence ratio     1
#> 82   anc_prevalence_age_matched           ANC prevalence age matched     1
#> 83 anc_art_coverage_age_matched         ANC ART coverage age matched     1
#>    accuracy format indicator_sort_order    iso3             colour min     max
#> 1        NA   0.0%                    2     MWI    interpolateReds   0 5.0e-01
#> 2        NA   0.0%                    2     MWI    interpolateReds   0 5.0e-01
#> 3        NA   0.0%                    2     MWI    interpolateReds   0 5.0e-01
#> 4        NA   0.0%                    2     MWI    interpolateReds   0 5.0e-01
#> 5        NA   0.0%                    2     MWI    interpolateReds   0 5.0e-01
#> 6        NA   0.0%                    4     MWI  interpolateGreens   0 1.0e+00
#> 7        NA   0.0%                    4     MWI  interpolateGreens   0 1.0e+00
#> 8        NA   0.0%                    4     MWI  interpolateGreens   0 1.0e+00
#> 9        NA   0.0%                    4     MWI  interpolateGreens   0 1.0e+00
#> 10       NA   0.0%                    4     MWI  interpolateGreens   0 1.0e+00
#> 11      100    0,0                    6     MWI  interpolateGreens   0 8.0e+04
#> 12      100    0,0                    6     MWI  interpolateGreens   0 8.0e+04
#> 13      100    0,0                   NA     MWI  interpolateGreens   0 8.0e+04
#> 14      100    0,0                    6     MWI  interpolateGreens   0 8.0e+04
#> 15      100    0,0                    6     MWI  interpolateGreens   0 8.0e+04
#> 16      100    0,0                   NA default  interpolateGreens   0 8.0e+04
#> 17      100    0,0                   NA default interpolatePurples   0 6.0e+04
#> 18      100    0,0                   NA default interpolatePurples   0 6.0e+04
#> 19        1  0.00%                   NA     MWI   interpolateBlues   0 6.0e-01
#> 20       NA   0.0%                   NA     MWI interpolatePurples   0 1.0e+00
#> 21       NA   0.0%                   NA     MWI interpolateOranges   0 5.0e-01
#> 22       NA   0.0%                   17     MWI interpolateOranges   0 5.0e-01
#> 23       NA   0.0%                   17     MWI interpolateOranges   0 5.0e-01
#> 24       NA   0.0%                   18     MWI interpolateOranges   0 1.0e+00
#> 25       NA   0.0%                   NA     MWI interpolateOranges   0 1.0e+00
#> 26       NA   0.0%                   18     MWI interpolateOranges   0 1.0e+00
#> 27      100    0,0                    5     MWI  interpolateGreens   0 7.0e+04
#> 28      100    0,0                    5     MWI  interpolateGreens   0 7.0e+04
#> 29     1000    0,0                    1     MWI   interpolateGreys   0 2.0e+07
#> 30     1000    0,0                    1     MWI   interpolateGreys   0 2.0e+07
#> 31     1000    0,0                    1     MWI   interpolateGreys   0 2.0e+07
#> 32      100    0,0                    3     MWI    interpolateReds   0 1.0e+07
#> 33      100    0,0                    3     MWI    interpolateReds   0 1.0e+07
#> 34      100    0,0                    3     MWI    interpolateReds   0 1.0e+07
#> 35       NA    0.0                   15     MWI   interpolateBlues   0 1.0e-02
#> 36       NA    0.0                   15     MWI   interpolateBlues   0 1.0e-02
#> 37       NA    0.0                   15     MWI   interpolateBlues   0 1.0e-02
#> 38      100    0,0                   16     MWI   interpolateBlues   0 4.0e+04
#> 39      100    0,0                   16     MWI   interpolateBlues   0 4.0e+04
#> 40      100    0,0                   16     MWI   interpolateBlues   0 4.0e+04
#> 41       NA   0.0%                    8 default   interpolateBlues   0 1.0e+00
#> 42       NA   0.0%                    8 default   interpolateBlues   0 1.0e+00
#> 43       NA   0.0%                    8 default   interpolateBlues   0 1.0e+00
#> 44       10      0                    9 default   interpolateBlues   0 8.0e+04
#> 45       10    0,0                    9 default   interpolateBlues   0 8.0e+04
#> 46       10    0,0                    9 default   interpolateBlues   0 8.0e+04
#> 47       10      0                   10 default   interpolateBlues   0 8.0e+04
#> 48       10    0,0                   10 default   interpolateBlues   0 8.0e+04
#> 49       10    0,0                    7 default  interpolateGreens   0 8.0e+04
#> 50       10    0,0                    7 default  interpolateGreens   0 8.0e+04
#> 51      100    0,0                   11 default    interpolateReds   0 1.0e+07
#> 52      100    0,0                   11 default    interpolateReds   0 1.0e+07
#> 53       10    0,0                   12 default  interpolateGreens   0 8.0e+04
#> 54       10    0,0                   12 default  interpolateGreens   0 8.0e+04
#> 55       10    0,0                   13 default   interpolateBlues   0 8.0e+04
#> 56       10    0,0                   13 default   interpolateBlues   0 8.0e+04
#> 57       10    0,0                   14 default   interpolateBlues   0 8.0e+04
#> 58       10    0,0                   14 default   interpolateBlues   0 8.0e+04
#> 59      100    0,0                   19 default interpolateOranges   0 7.0e+05
#> 60      100    0,0                   19 default interpolateOranges   0 7.0e+05
#> 61       10    0,0                   20 default interpolateOranges   0 3.5e+05
#> 62       10    0,0                   20 default interpolateOranges   0 3.5e+05
#> 63       10    0,0                   21 default interpolateOranges   0 3.5e+05
#> 64       10    0,0                   21 default interpolateOranges   0 3.5e+05
#> 65       10    0,0                   22 default interpolateOranges   0 3.5e+05
#> 66       10    0,0                   22 default interpolateOranges   0 3.5e+05
#> 67       10    0,0                   23 default interpolateOranges   0 3.5e+05
#> 68       10    0,0                   23 default interpolateOranges   0 3.5e+05
#> 69       10    0,0                   24 default interpolateOranges   0 3.5e+05
#> 70       10    0,0                   24 default interpolateOranges   0 3.5e+05
#> 71       10    0,0                   25 default interpolateOranges   0 7.0e+05
#> 72       10    0,0                   25 default interpolateOranges   0 7.0e+05
#> 73       NA    0.0                   26 default   interpolateGreys   0 1.2e+00
#> 74       NA    0.0                   27 default    interpolateReds   0 1.2e+00
#> 75       NA    0.0                   28 default   interpolateBlues   0 1.2e+00
#> 76       NA    0.0                   29 default  interpolateGreens   0 1.2e+00
#> 77       NA    0.0                   30 default   interpolateBlues   0 1.2e+00
#> 78       NA    0.0                   31 default    interpolateReds   0 1.2e+00
#> 79       NA    0.0                   32 default  interpolateGreens   0 1.2e+00
#> 80       NA    0.0                   33 default   interpolateBlues   0 1.2e+00
#> 81       NA    0.0                   34 default   interpolateBlues   0 1.2e+00
#> 82       NA   0.0%                   35 default interpolateOranges   0 5.0e-01
#> 83       NA   0.0%                   36 default interpolateOranges   0 1.0e+00
#>    invert_scale
#> 1         FALSE
#> 2         FALSE
#> 3         FALSE
#> 4         FALSE
#> 5         FALSE
#> 6         FALSE
#> 7         FALSE
#> 8         FALSE
#> 9         FALSE
#> 10        FALSE
#> 11        FALSE
#> 12        FALSE
#> 13        FALSE
#> 14        FALSE
#> 15        FALSE
#> 16        FALSE
#> 17        FALSE
#> 18        FALSE
#> 19        FALSE
#> 20        FALSE
#> 21        FALSE
#> 22        FALSE
#> 23        FALSE
#> 24        FALSE
#> 25        FALSE
#> 26        FALSE
#> 27        FALSE
#> 28        FALSE
#> 29        FALSE
#> 30        FALSE
#> 31        FALSE
#> 32        FALSE
#> 33        FALSE
#> 34        FALSE
#> 35        FALSE
#> 36        FALSE
#> 37        FALSE
#> 38        FALSE
#> 39        FALSE
#> 40        FALSE
#> 41        FALSE
#> 42        FALSE
#> 43        FALSE
#> 44        FALSE
#> 45        FALSE
#> 46        FALSE
#> 47        FALSE
#> 48        FALSE
#> 49        FALSE
#> 50        FALSE
#> 51        FALSE
#> 52        FALSE
#> 53        FALSE
#> 54        FALSE
#> 55        FALSE
#> 56        FALSE
#> 57        FALSE
#> 58        FALSE
#> 59        FALSE
#> 60        FALSE
#> 61        FALSE
#> 62        FALSE
#> 63        FALSE
#> 64        FALSE
#> 65        FALSE
#> 66        FALSE
#> 67        FALSE
#> 68        FALSE
#> 69        FALSE
#> 70        FALSE
#> 71        FALSE
#> 72        FALSE
#> 73        FALSE
#> 74        FALSE
#> 75        FALSE
#> 76        FALSE
#> 77        FALSE
#> 78        FALSE
#> 79        FALSE
#> 80        FALSE
#> 81        FALSE
#> 82        FALSE
#> 83        FALSE