Construct Model Frames and Adjacency Structures

Construct Model Frames and Adjacency Structures

naomi_model_frame(
  area_merged,
  population_agesex,
  spec,
  scope = area_merged$area_id[area_merged$area_level == min(area_merged$area_level)],
  level = max(area_merged$area_level),
  calendar_quarter1,
  calendar_quarter2,
  calendar_quarter3,
  calendar_quarter4 = "CY2024Q3",
  calendar_quarter5 = "CY2025Q3",
  age_groups = get_five_year_age_groups(),
  sexes = c("male", "female"),
  omega = 0.7,
  rita_param = list(OmegaT0 = 130/365, sigma_OmegaT = ((142 - 118)/365)/(2 *
    stats::qnorm(0.975)), betaT0 = 0, sigma_betaT = 1e-05, ritaT = 1),
  sigma_u_sd = 1,
  artattend = TRUE,
  artattend_t2 = FALSE,
  artattend_log_gamma_offset = -4,
  anchor_home_district = TRUE,
  rho_paed_15to49f_ratio = TRUE,
  rho_paed_x_term = FALSE,
  alpha_xst_term = FALSE,
  logit_nu_mean = 2,
  logit_nu_sd = 0.3,
  spectrum_population_calibration = "national",
  output_aware_plhiv = TRUE,
  adjust_area_growth = FALSE,
  psnu_level = NULL
)

Arguments

area_merged

Merged version of area hierarchy

population_agesex

Population by age group and sex

spec

Spec

scope

The collection of area IDs to be modelled. Defaults to all area ids.

level

Admin level

calendar_quarter1

Calendar quarter at time 1 ("CYyyyyQq")

calendar_quarter2

Calendar quarter at time 2 ("CYyyyyQq")

calendar_quarter3

Calendar quarter at time 3 ("CYyyyyQq")

calendar_quarter4

Calendar quarter at time 4 ("CYyyyyQq")

calendar_quarter5

Calendar quarter at time 5 ("CYyyyyQq")

age_groups

Age groups to include in model frame

sexes

Sexes

omega

Omega

rita_param

rita_param

sigma_u_sd

sigma_u_sd

artattend

logical; whether to estimate neighboring district ART attendance

artattend_t2

logical; whether to allow time-varying neighboring district ART attendance

artattend_log_gamma_offset

logit offset for neighboring district ART attendance

anchor_home_district

logical; whether to include home district random in ART attendance specification

rho_paed_15to49f_ratio

logical; to model paediatric prevalence as ratio of 15-49 female prevalence

rho_paed_x_term

logical; to include area interaction for paediatric prevalence

alpha_xst_term

logical; to include district-sex-time interaction for ART coverage. Default FALSE.

logit_nu_mean

mean of logit viral load suppression.

logit_nu_sd

standard deviation of logit viral load suppression.

spectrum_population_calibration

character string values "national", "subnational", "none"

adjust_area_growth

TRUE/FALSE: adjust PLHIV population for net change in cohort size

psnu_level

Level at which to output PEPFAR Data Pack outputs for planning. This doesn't have anything to do with calibration, but this is a convenient place to allow user to edit this in model workflow. Default value NULL indicates to read PSNU level from internal database.

Value

Naomi model frame

Details

Argument spectrum_population_calibration determines whether to calibrate population inputs to match Spectrum population by age and sex. If the Spectrum file is a single national Spectrum file, then options "national" and "subnational" return the same results.